GAM functions for Xpose 4

These are functions for summarizing and plotting the results of the generalized additive model within Xpose

Usage

xp.akaike.plot(
  gamobj = NULL,
  title = "Default",
  xlb = "Akaike value",
  ylb = "Models",
  ...
)

xp.cook(gam.object)

xp.ind.inf.fit(
  gamobj = NULL,
  plot.ids = TRUE,
  idscex = 0.7,
  ptscex = 0.7,
  title = "Default",
  recur = FALSE,
  xlb = NULL,
  ylb = NULL,
  ...
)

xp.ind.inf.terms(
  gamobj = NULL,
  xlb = NULL,
  ylb = NULL,
  plot.ids = TRUE,
  idscex = 0.7,
  ptscex = 0.7,
  prompt = TRUE,
  ...
)

xp.ind.stud.res(
  gamobj = NULL,
  title = "Default",
  recur = FALSE,
  xlb = NULL,
  ylb = NULL
)

xp.plot(
  gamobj = NULL,
  plot.ids = TRUE,
  idscex = 0.7,
  ptscex = 0.7,
  prompt = TRUE,
  ...
)

xp.summary(gamobj = NULL)

Arguments

gamobj

A GAM object to use in the plot. IF null then the user is asked to choose from a list of GAM objects in memory.

title

A text string indicating plot title. If NULL, left blank.

xlb

A text string indicating x-axis legend. If NULL, left blank.

ylb

A text string indicating y-axis legend. If NULL, left blank.

...

Other arguments passed to the GAM functions.

gam.object

A GAM object (see gam.

plot.ids

Logical, specifies whether or not ID numbers should be displayed.

idscex

ID label size.

ptscex

Point size.

recur

If dispersion should be used in the GAM object.

prompt

Specifies whether or not the user should be prompted to press RETURN between plot pages. Default is TRUE.

object

An xpose.data object.

Value

Plots or summaries.

Functions

xp.akaike.plot(): An Akaike plot of the results.

xp.cook(): Individual parameters to GAM fit.

xp.ind.inf.fit(): Individual influence on GAM fit.

xp.ind.inf.terms(): Individual influence on GAM terms.

xp.ind.stud.res(): Studentized residuals.

xp.plot(): GAM residuals of base model vs. covariates.

xp.summary(): Summarize GAM.

See also

Author

Niclas Jonsson & Andrew Hooker