absolute value of the individual weighted residuals vs. the independent variable

absolute value of the individual weighted residuals vs. the independent variable

absval.iwres.vs.idv(
  object,
  ylb = "|iWRES|",
  smooth = TRUE,
  idsdir = "up",
  type = "p",
  ...
)

Arguments

object
An "xpose.data" object.

ylb
A string giving the label for the y-axis. NULL if none.

smooth
A NULL value indicates that no superposed line should be added to the graph. If TRUE then a smooth of the data will be superimposed.

idsdir
a string indicating the directions of the extremes to include in labelling. Possible values are "up", "down" and "both".

type
1-character string giving the type of plot desired. The following values are possible, for details, see 'plot': '"p"' for points, '"l"' for lines, '"o"' for over-plotted points and lines, '"b"', '"c"') for (empty if '"c"') points joined by lines, '"s"' and '"S"' for stair steps and '"h"' for histogram-like vertical lines. Finally, '"n"' does not produce any points or lines.

...
Other arguments passed to xpose.panel.default.

Value

A lattice object

See also