Generic internal functions for Xpose 4

These are internal functions relating to the Xpose generic functions.

addid(x, y, ids = ids, idsmode = NULL, idsext = 0.05, idscex = 0.7,
  idsdir = "both", gridmode = TRUE)

computePI(x, y, object, limits = object@Prefs@Graph.prefs$PIlimits,
  logy = FALSE, logx = FALSE, onlyfirst = FALSE, inclZeroWRES = FALSE,
  PI.subset = NULL, subscripts)

create.rand(data, object, frac, seed = NULL)

create.strat.rand(data, object, x, y, frac, dilci, seed = NULL)

eq.xpose(x, number = 6, overlap = 0.5)

get.refrunno(database = ".ref.db")

xpose.stack(data, object, select, rep, subset = NULL, ...)

Value

Internal helper functions for the generic Xpose functions.

Details

These are internal Xpose functions, for adding ID numbers, computing prediction intervals, randomization, stacking, and binning. They are not intended for direct use.