Basic model comparison plots, for Xpose 4

This creates a stack of four plots, comparing PRED, IPRED, WRES (or CWRES), and IWRES estimates for the two specified model fits.

basic.model.comp(object, object.ref = NULL, onlyfirst = FALSE,
  inclZeroWRES = FALSE, subset = xsubset(object), main = "Default",
  force.wres = FALSE, ...)

Arguments

object
An xpose.data object.

object.ref
An xpose.data object. If not supplied, the user will be prompted.

onlyfirst
Logical value indicating whether only the first row per individual is included in the plot.

inclZeroWRES
Logical value indicating whether rows with WRES=0 is included in the plot. The default is TRUE.

subset
A string giving the subset expression to be applied to the data before plotting. See xsubset.

main
The title of the plot. If "Default" then a default title is plotted. Otherwise the value should be a string like "my title" or NULL for no plot title.

force.wres
Force function to use WRES?

…
Other arguments passed to link{xpose.plot.default}.

Value

Returns a stack of plots comprising comparisons of PRED, IPRED, WRES (or CWRES) and IWRES for the two specified runs.

Details

Four basic model comparison plots are displayed in sequence.

Conditional weighted residuals (CWRES) require some extra steps to calculate. See compute.cwres for details.

A wide array of extra options controlling xyplots are available. See xpose.plot.default for details.

See also

xpose.plot.default, xpose.panel.default, xyplot, compute.cwres, xpose.prefs-class, xpose.data-class

Other specific functions: absval.cwres.vs.cov.bw, absval.cwres.vs.pred.by.cov, absval.cwres.vs.pred, absval.iwres.cwres.vs.ipred.pred, absval.iwres.vs.cov.bw, absval.iwres.vs.idv, absval.iwres.vs.ipred.by.cov, absval.iwres.vs.ipred, absval.iwres.vs.pred, absval.wres.vs.cov.bw, absval.wres.vs.idv, absval.wres.vs.pred.by.cov, absval.wres.vs.pred, absval_delta_vs_cov_model_comp, addit.gof, autocorr.cwres, autocorr.iwres, autocorr.wres, basic.gof, cat.dv.vs.idv.sb, cat.pc, cov.splom, cwres.dist.hist, cwres.dist.qq, cwres.vs.cov, cwres.vs.idv.bw, cwres.vs.idv, cwres.vs.pred.bw, cwres.vs.pred, cwres_wres_vs_x, dOFV.vs.cov, dOFV.vs.id, dOFV1.vs.dOFV2, data.checkout, dv.preds.vs.idv, dv.vs.idv, dv.vs.ipred.by.cov, dv.vs.ipred.by.idv, dv.vs.ipred, dv.vs.pred.by.cov, dv.vs.pred.by.idv, dv.vs.pred.ipred, dv.vs.pred, gof, ind.plots.cwres.hist, ind.plots.cwres.qq, ind.plots, ipred.vs.idv, iwres.dist.hist, iwres.dist.qq, iwres.vs.idv, kaplan.plot, par_cov_hist, par_cov_qq, parm.vs.cov, parm.vs.parm, pred.vs.idv, ranpar.vs.cov, runsum, wres.dist.hist, wres.dist.qq, wres.vs.idv.bw, wres.vs.idv, wres.vs.pred.bw, wres.vs.pred, xpose.VPC.both, xpose.VPC.categorical, xpose.VPC, xpose4-package

Examples

not_run({
  ## We expect to find the required NONMEM run and table files for runs
  ## 5 and 6 in the current working directory
  xpdb5 <- xpose.data(5)
  xpdb6 <- xpose.data(6)

  ## A vanilla plot, without prompts
  basic.model.comp(xpdb5, xpdb6, prompt = FALSE)

  ## Custom colours and symbols, no user IDs
  basic.model.comp.cwres(xpdb5, xpdb6, cex=0.6, pch=8, col=1, ids=NULL)
})