Function to create histograms of results from the bootstrap tool in PsN

Reads results from the bootstrap tool in PsN and then creates histograms.

boot.hist(
  results.file = "raw_results_run1.csv",
  incl.ids.file = "included_individuals1.csv",
  min.failed = FALSE,
  cov.failed = FALSE,
  cov.warnings = FALSE,
  boundary = FALSE,
  showOriginal = TRUE,
  showMean = FALSE,
  showMedian = FALSE,
  showPCTS = FALSE,
  PCTS = c(0.025, 0.975),
  excl.id = c(),
  layout = NULL,
  sort.plots = TRUE,
  main = "Default",
  ...
)

Arguments

results.file
The location of the results file from the bootstrap tool in PsN

incl.ids.file
The location of the included ids file from the bootstrap tool in PsN

min.failed
Should NONMEM runs that had failed minimization be skipped? TRUE or FALSE

cov.failed
Should NONMEM runs that had a failed covariance step be skipped? TRUE or FALSE

cov.warnings
Should NONMEM runs that had covariance step warnings be skipped? TRUE or FALSE

boundary
Should NONMEM runs that had boundary warnings be skipped? TRUE or FALSE

showOriginal
Should we show the value from the original NONMEM run in the histograms? TRUE or FALSE

showMean
Should we show the mean of the histogram data? TRUE or FALSE

showMedian
Should we show the median of the histogram data? TRUE or FALSE

showPCTS
Should we show the percentiles of the histogram data? TRUE or FALSE

PCTS
the percentiles to show. Can be a vector of any length. For example, c(0.05,0.2,0.5,0.7)

excl.id
Vector of id numbers to exclude.

layout
Layout of plots. A vector of number of rows and columns in each plot. c(3,3) for example.

sort.plots
Should the plots be sorted based on type of parameter. Sorting on parameters, standard errors, shrinkage and eigenvalues.

main
The title of the plot.

...
Additional arguments that can be passed to xpose.plot.histogram, xpose.panel.histogram, histogram and other lattice-package functions.

Value

A lattice object

References

PsN

See also

Author

Andrew Hooker

Examples

if (FALSE) {
boot.hist(results.file="./boot1/raw_results_run1.csv", 
          incl.ids.file="./boot1/included_individuals1.csv")
}