This function creates the Xpose data classes ("xpose.data" and "xpose.prefs")

This function defines and sets the Xpose data classes.

createXposeClasses(nm7 = F)

Arguments

nm7

FALSE if not using NONMEM 7.

Note

All the default settings are defined in this function.

See also

Author

Niclas Jonsson and Andrew C. Hooker