Exports Xpose variable definitions to a file from an Xpose data object.

This function exports variable definitions for a specified Xpose data object to a file.

export.variable.definitions(object, file = "")

Arguments

object
An xpose.data object.

file
A file name as a string.

Value

Null.

Details

This function exports variable defintions (contents of object@Prefs@Xvardef) for a given xpose.data object to a file, typically 'xpose.vardefs.ini'. Note that file format is not the same as used for graphics settings. It is a wrapper for the R function dput.

See also

import.variable.definitions, xpose.prefs-class dput

Other data functions: add_transformed_columns, change_graphical_parameters, change_misc_parameters, compute.cwres, data.checkout, data_extract_or_assign, db.names, export.graph.par, import.graph.par, import.variable.definitions, make.sb.data, nsim, par_cov_summary, read.TTE.sim.data, read.nm.tables, read_NM_output, read_nm_table, simprazExample, tabulate.parameters, xlabel, xpose.data, xpose.print, xpose4-package, xsubset

Examples

od = setwd(tempdir()) # move to a temp directory
(cur.files <- dir()) # current files in temp directory

#>  [1] "00f3ba3af1e746448d26df9e1e60660b.png"                       
#>  [2] "0638b8a3768149fe918720e5c2ef18e9.png"                       
#>  [3] "0b67319d692742c8834591162a3647ad.png"                       
#>  [4] "11f19707dbfd472e8ba94d3768f7aeeb.png"                       
#>  [5] "197d0522dbcd4ad688ca8d2fecc499ba.png"                       
#>  [6] "1c92af1409744895835a3c0d4e2c59f6.png"                       
#>  [7] "20ba4ccd7924404d8413e0b2d0ae403a.png"                       
#>  [8] "21757f8c08214490ad6998846c76127a.png"                       
#>  [9] "25b2bb56152d44b9a70b189bfc26c08d.png"                       
#> [10] "28159ee4a1f24c95b79d02665774ec23.png"                       
#> [11] "28feb19f05f74b2f8a0f506418bb19ef.png"                       
#> [12] "2ca138e8c45b4f49bb522521fdc77362.png"                       
#> [13] "2f28fd2d0501473bb17a6290901ca780.png"                       
#> [14] "35565301aab54235a5e91cc4d8f8848a.png"                       
#> [15] "3695e46e124e4a8dbd8b773dec150b43.png"                       
#> [16] "3ff4926f0722486ca9ff4795a84877d6.png"                       
#> [17] "401a2ab6d2fd4d65a7775ce010fb0932.png"                       
#> [18] "403eca9f85424f18b65e3c7356ef7dc6.png"                       
#> [19] "42bc132d787f43278d56ce679e15ac86.png"                       
#> [20] "45cf4a2d43a34bfcbc57deb4b54f5628.png"                       
#> [21] "4aab54725b1f4f9384cedf2de19fa2bc.png"                       
#> [22] "4db642e580bf4f8e8086765d3a676288.png"                       
#> [23] "581fff3a38984d52bdbd385611cebf4e.png"                       
#> [24] "59f44cddfb284bc18a4b102eb361cf01.png"                       
#> [25] "5d20eaffca9445ef8334872301f44db1.png"                       
#> [26] "5e3c7971491044f7a79e06d69ebb84cc.png"                       
#> [27] "5ebdec48be8b4576a60afe3881ddeac6.png"                       
#> [28] "6719736f41f645498654ece7d6047721.png"                       
#> [29] "68033ae12b0d49e0b42238945b34c82c.png"                       
#> [30] "6bec1bdac7974acc988405b58e435aab.png"                       
#> [31] "6f9c5812f4964d71b1049eda02ad0bc9.png"                       
#> [32] "701efb73cdc3480c9a43ce8e61b33f62.png"                       
#> [33] "70cad80522df4008be88e80e8626692e.png"                       
#> [34] "712e381ea07a49358cdc9598c477500a.png"                       
#> [35] "7b865e52c8c74aa99fec9442c673a979.png"                       
#> [36] "7de295ca47944af8bea81d20c910465c.png"                       
#> [37] "7f536c6e88e04011a7f17b0cd093392b.png"                       
#> [38] "81efcf41ff4d4a54a549f66edc699598.png"                       
#> [39] "878fe077c41f41ddb2ce6307f8668572.png"                       
#> [40] "881e5065206e4be088238cd1b3423ea1.png"                       
#> [41] "88bf43dc17064cca9b9180fbf8260ef7.png"                       
#> [42] "8d1cb2598c834ea2946c5ba72702d82b.png"                       
#> [43] "90e3ca4a34444c9991917ea2276336da.png"                       
#> [44] "9165c672dcdc4b0b98d2138557dc7c59.png"                       
#> [45] "93bf0f7607f144f8961edd2fe2576d5e.png"                       
#> [46] "94b337aa422d4474a06c35c8c477c3c9.png"                       
#> [47] "993e507847544181a76681b2615cdff4.png"                       
#> [48] "9b111c4d753e46c0a912975df1291073.png"                       
#> [49] "9f5acd001bc14575af18627411eacee4.png"                       
#> [50] "a36c5bcd24b9443e949c0b2f321747ca.png"                       
#> [51] "a8fab28f868146919f2079cebd675b4b.png"                       
#> [52] "b3d75c5614ab4674afc70c7adbd13fe0.png"                       
#> [53] "b8f9fc39316242c09b9d54f3a8f69a50.png"                       
#> [54] "b9525313161745a2be0cf43c4711e8ab.png"                       
#> [55] "ba2ed6adc93f49c4a380fa681807e607.png"                       
#> [56] "ba4a3bee1fe245f18927d7758b0fce88.png"                       
#> [57] "bc4fcdd8024c4e69b502d2d4b83b9146.png"                       
#> [58] "bc8958eb5fc14689ba77a3e99c2811bf.png"                       
#> [59] "bd8f047874174c4c86b6cecb262165e8.png"                       
#> [60] "c4ca23a1628b48eb90957ce719539411.png"                       
#> [61] "c714bcab2fad4bb1bbe83100018a4fe6.png"                       
#> [62] "cebcdd15195a453fbc447b68db923814.png"                       
#> [63] "d3fd48fc17514e44ad17d037f76890cb.png"                       
#> [64] "d9c87dc69f524f47b6bb771ff7bcb63c.png"                       
#> [65] "da32570d635c4cb5b413e735068d3769.png"                       
#> [66] "dd91b6971df0441aa131201c7113544d.png"                       
#> [67] "e0a7fcc164aa4a3f8bdaa10c56b63ca3.png"                       
#> [68] "e4061249203c4e94a13d34236184b132.png"                       
#> [69] "e7f1fbad4eba491d9c8bb633ecc2681c.png"                       
#> [70] "e8ed7cbccaa44733a01800619b10df95.png"                       
#> [71] "e8fe3847598949f8bab113a4ada1295a.png"                       
#> [72] "eb699b1fd5834d738f72609a845fdd28.png"                       
#> [73] "f10f138210d64bf49d9b004a9100f96e.png"                       
#> [74] "f2e612bea1564ce9a36f1759d9eee964.png"                       
#> [75] "f6e123e4ac8f469f912cbf60cf2334e3.png"                       
#> [76] "gitmsge8d495c6843.txt"                                      
#> [77] "gitmsge8d55c20d1a.txt"                                      
#> [78] "libloc_213_8f8c4eafe9e43f75.rds"                            
#> [79] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"

export.variable.definitions(simpraz.xpdb,file="xpose.vardefs.ini")
(new.files <- dir()[!(dir() %in% cur.files)])  # what files are new here?

#> [1] "xpose.vardefs.ini"

file.remove(new.files) # remove this file

#> [1] TRUE

setwd(od)  # restore working directory