Imports Xpose graphics settings from a file to an Xpose data object.

This function imports graphics settings for a specified Xpose data object from a file.

Usage

import.graph.par(object, classic = FALSE)

Arguments

object

An xpose.data object.

classic

A logical operator specifying whether the function should assume the classic menu system. This is an internal option and need never be called from the command line.

Value

An xpose.data object (classic = FALSE) or null (classic = TRUE).

Details

This function imports graphics settings (contents of object@Prefs@Graph.prefs) for a given xpose.data object from a file, typically 'xpose.ini'. It is a wrapper for xpose.read. It returns an xpose.data object. Note that the file format is not the same as is used in import.variable.definitions and export.variable.definitions.

See also

Author

Niclas Jonsson & Justin Wilkins

Examples

if (FALSE) { # \dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Import graphics preferences you saved earlier using export.graph.par
xpdb5 <- import.graph.par(xpdb5)

## Command-line driven
xpdb5 <- xpose.read(xpdb5, "c:/XposeSettings/mytheme.ini")
} # }