Function to plot the coverage of the Numerical Predictive Check

This function takes the output from the npc command in Perl Speaks NONMEM (PsN) and makes a coverage plot. A coverage plot for the NPC looks at different prediction intervals (PIs) for each data point and calculates the total number of data points in the data set lying outside of these PIs. The plot shows the relative amount of data points outside of their PI compared to the expected amount at that PI. In addition a confidence interval around these values are computed based on the simulated data.

npc.coverage(
  npc.info = "npc_results.csv",
  main = "Default",
  main.sub = NULL,
  main.sub.cex = 0.85,
  ...
)

Arguments

npc.info
The results file from the npc command in PsN. for example npc\_results.csv, or if the file is in a separate directory ./npc\_dir1/npc\_results.csv.

main
A string giving the plot title or NULL if none. "Default" creates a default title.

main.sub
Used for names above each plot when using multiple plots. Should be a vector c("Group 1","Group 2")

main.sub.cex
The size of the main.sub titles.

...
Other arguments passed to xpose.multiple.plot.default, xyplot and others. Please see these functions for more descriptions of what you can do.

Value

A list of plots

Additional arguments for the NPC coverage plots

Additional plot features

list("CI = \"both\", \"area\" or \"lines\"")

Specifies whether confidence intervals (as lines, a shaded area or both) should be added to the plot. NULL means no CI.

list("mark.outside.data = TRUE or FALSE ")

Should the points outside the CI be marked in a different color to identify them.

list("abline = TRUE")

Should there be a line to mark the value of y=1? Possible values are TRUE, FALSE and NULL.

Should there be a line to mark the value of y=1? Possible values are TRUE, FALSE and NULL.

Line and area control. See plot, grid.polygon and xyplot for more details.

list("CI.area.col = \"blue\"")

What color should the area for the CI be? Defaults to "blue".

list("CI.area.alpha = 0.3")

How much transparency should the CI.area.col have? Defaults to 0.3.

list("ab.lwd=1")

The width of the abline.

list("ab.lty=\"dashed\"")

How should the abline look?

list("CI.upper.lty=\"dotted\"")

What should the line at the upper edge of the CI look like when using CI = "both" or "lines"?

What should the line at the upper edge of the CI look like when using CI = "both" or "lines"?

list("CI.upper.col=\"brown\"")

What color should the line at the upper edge of the CI have when using CI = "both" or "lines"?

list("CI.upper.lwd=\"2\" ")

The line width of the line at the upper edge of the CI when using CI = "both" or "lines"?

list("CI.lower.lty=\"dotted\"")

What should the line at the lower edge of the CI look like when using CI = "both" or "lines"?

What should the line at the lower edge of the CI look like when using CI = "both" or "lines"?

list("CI.lower.col=\"brown\"")

What color should the line at the lower edge of the CI have when using CI = "both" or "lines"?

list("CI.lower.lwd=\"2\" ")

The line width of the line at the lower edge of the CI when using CI = "both" or "lines"?

list("obs.col=\"black\"")

The color of the observed values.

list("obs.pch=19")

The type of point to use for the observed values.

list("obs.lty=\"solid\"")

The type of line to use for the observed values.

list("obs.type=\"b\"")

The combination of lines and points to use for the observed values.

list("obs.cex=1")

The size of the points to use for the observed values.

list("obs.lwd=1")

The line width to use for the observed values.

list("out.col=\"red\"")

The color of the observed values that lie outside of the CI. Only used if mark.outside.data = TRUE.

list("out.pch=8")

The type of point to use for the observed values that lie outside of the CI. Only used if mark.outside.data = TRUE.

list("out.cex=1.3")

The size of the points of the observed values that lie outside of the CI. Only used if mark.outside.data = TRUE.

list("out.lwd=1")

The line width of the observed values that lie outside of the CI. Only used if mark.outside.data = TRUE.

See also

Examples

if (FALSE) {
library(xpose4)

npc.coverage()

## to read files in a directory different than the current working directory 
npc.file <- "./another_directory/npc_results.csv"
npc.coverage(npc.info=npc.file)
}