Parameters plotted against covariates, for Xpose 4

This creates a stack of plots of Bayesian parameter estimates plotted against covariates, and is a specific function in Xpose 4. It is a wrapper encapsulating arguments to the xpose.plot.default function. Most of the options take their default values from xpose.data object but may be overridden by supplying them as arguments.

Usage

parm.vs.cov(
  object,
  onlyfirst = TRUE,
  smooth = TRUE,
  type = "p",
  main = "Default",
  ...
)

Arguments

object

An xpose.data object.

onlyfirst

Logical value indicating whether only the first row per individual is included in the plot.

smooth

Logical value indicating whether an x-y smooth should be superimposed. The default is TRUE.

type

The plot type - defaults to points only.

main

The title of the plot. If "Default" then a default title is plotted. Otherwise the value should be a string like "my title" or NULL for no plot title.

...

Other arguments passed to link{xpose.plot.default}.

Value

Returns a stack of xyplots and histograms of parameters against covariates.

Details

Each of the parameters in the Xpose data object, as specified in object@Prefs@Xvardef$parms, is plotted against each covariate present, as specified in object@Prefs@Xvardef$covariates, creating a stack of plots.

A wide array of extra options controlling xyplots are available. See xpose.plot.default and xpose.panel.default for details.

See also

Author

E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins

Examples

if (FALSE) {
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb <- xpose.data(5)

## A vanilla plot
parm.vs.cov(xpdb)

## Custom colours and symbols, IDs
parm.vs.cov(xpdb, cex=0.6, pch=3, col=1, ids=TRUE)
}