Read the vpctab file from PsN into Xpose

This function read in the vpctab file created from PsN and gathers the information needed to make a vpc plot.

read.vpctab(vpctab = NULL, object = NULL, vpc.name = "vpctab",
  vpc.suffix = "", tab.suffix = "", inclZeroWRES = FALSE,
  verbose = FALSE, ...)

Arguments

vpctab
The vpctab file from a 'vpc' run in PsN.

object
An xpose data object. Created from xpose.data. One of object or vpctab is required. If both are present then the information from the vpctab will over-ride the xpose data object object (i.e. the values from the vpctab will replace any matching values in the object@Data portion of the xpose data object). If only object is present then the function will look for a vpctab with the same run number as the one associated with the object.

vpc.name
The default name of the vpctab file. Used if only object is supplied.

vpc.suffix
The suffix of the vpctab file. Used if only object is supplied.

tab.suffix
The table suffix of the vpctab file. Used if only object is supplied. Final order of the file would be then paste(vpc.name,object@Runno,vpc.suffix,tab.suffix)

inclZeroWRES
If there are no zero valued weighted residuals in the object then this should be TRUE.

verbose
Text messages passed to screen or not.

…
Other arguments passed to other functions.

Value

Returned is an xpose data object with vpctab information included.

See also

xpose.VPC

Other PsN functions: boot.hist, bootscm.import, npc.coverage, randtest.hist, read.npc.vpc.results, xpose.VPC.both, xpose.VPC.categorical, xpose.VPC, xpose4-package