Simulated prazosin Xpose database.

Xpose database from the NONMEM output of a model for prasozin using simulated data (and NONMEM 7.3).

simpraz.xpdb

Format

an xpose.data object

Details

The database can be used to test functions in Xpose 4. This database is slightly different than the database that is created when reading in the files created by simprazExample using xpose.data.

See also

simprazExample

Examples

xpose.print(simpraz.xpdb)

#> The database contains the following observed items:
#> ID TIME IPRED IWRES CWRES CL V KA ETA1 ETA2 ETA3 AGE HT WT 
#> SECR SEX RACE SMOK HCTZ PROP CON OCC DV PRED RES WRES
#> 
#> The following variables are defined:
#> 
#> ID variable: ID 
#> Label variable: ID 
#> Independent variable: TIME 
#> Occasion variable: OCC 
#> Dependent variable: DV 
#> Population prediction variable: PRED 
#> Individual prediction variable: IPRED 
#> Weighted population residual variable: WRES 
#> Weighted individual residual variable: IWRES 
#> Population residual variable: RES 
#> Parameters: ETA3 ETA2 ETA1 KA V CL
#> Covariates: SEX RACE SMOK HCTZ PROP CON OCC AGE HT WT SECR
#>   ( Continuous: AGE HT WT SECR )
#>   ( Categorical: SEX RACE SMOK HCTZ PROP CON OCC )
#> Variability parameters: ETA1 ETA2 ETA3
#> Missing value label: -99 

Data(simpraz.xpdb)

#>     ID TIME      IPRED       IWRES      CWRES      CL       V      KA
#> 2    1    1 6.9193e+01  0.03681300 -0.0646470 13.5790  93.640 1.22490
#> 3    1    2 8.0181e+01 -0.09442400 -0.9411300 13.5790  93.640 1.22490
#> 4    1    3 7.5330e+01  0.16833000  1.1911000 13.5790  93.640 1.22490
#> 5    1    4 6.6916e+01 -0.20602000 -1.5154000 13.5790  93.640 1.22490
#> 6    1    5 5.8398e+01 -0.02685300 -0.0596360 13.5790  93.640 1.22490
#> 7    1    6 5.0666e+01  0.02513800  0.4099200 13.5790  93.640 1.22490
#> 8    1    7 4.3871e+01  0.20557000  1.8304000 13.5790  93.640 1.22490
#> 9    1    9 3.2841e+01 -0.17939000 -1.0219000 13.5790  93.640 1.22490
#> 10   1   11 2.4575e+01  0.06492000  0.8373000 13.5790  93.640 1.22490
#> 12   2    1 9.6073e+01  0.13195000  0.8119200  8.7284  92.501 3.01420
#> 13   2    2 9.2138e+01  0.04842400 -0.0866680  8.7284  92.501 3.01420
#> 14   2    3 8.4073e+01 -0.03655400 -0.5405500  8.7284  92.501 3.01420
#> 15   2    4 7.6514e+01  0.00726370  0.0058890  8.7284  92.501 3.01420
#> 16   2    5 6.9625e+01 -0.07260400 -0.4024500  8.7284  92.501 3.01420
#> 17   2    6 6.3356e+01 -0.20749000 -1.2486000  8.7284  92.501 3.01420
#> 18   2    7 5.7651e+01  0.12019000  1.3843000  8.7284  92.501 3.01420
#> 19   2    8 5.2460e+01 -0.03659500  0.3323700  8.7284  92.501 3.01420
#> 20   2   10 4.3438e+01 -0.04322700  0.4971900  8.7284  92.501 3.01420
#> 21   2   12 3.5967e+01  0.18052000  2.3421000  8.7284  92.501 3.01420
#> 23   3    1 8.7842e+00  0.06440800  0.3927400 11.0240  93.943 2.23050
#> 24   3    2 8.7557e+00 -0.01093200 -0.4033600 11.0240  93.943 2.23050
#> 25   3    3 7.8877e+00  0.03705600  0.0613180 11.0240  93.943 2.23050
#> 26   3    4 7.0253e+00 -0.11889000 -0.9633300 11.0240  93.943 2.23050
#> 27   3    5 6.2486e+00  0.13306000  1.0866000 11.0240  93.943 2.23050
#> 28   3    6 5.5568e+00 -0.15779000 -0.9918700 11.0240  93.943 2.23050
#> 29   3    7 4.9416e+00  0.07253700  0.8514900 11.0240  93.943 2.23050
#> 30   3    8 4.3944e+00 -0.04423900  0.0490350 11.0240  93.943 2.23050
#> 31   3   10 3.4751e+00  0.12801000  1.4633000 11.0240  93.943 2.23050
#> 32   3   12 2.7482e+00  0.00066303  0.5552500 11.0240  93.943 2.23050
#> 34   4    1 6.3799e+01 -0.01017600  0.4475200 19.6060  49.991 1.67760
#> 35   4    2 5.5019e+01  0.13433000  1.2201000 19.6060  49.991 1.67760
#> 36   4    3 3.9396e+01  0.04198600  0.2016500 19.6060  49.991 1.67760
#> 37   4    4 2.7031e+01 -0.10065000 -1.0776000 19.6060  49.991 1.67760
#> 38   4    5 1.8339e+01 -0.00920370 -0.4784300 19.6060  49.991 1.67760
#> 39   4    6 1.2404e+01  0.06662200  0.0711500 19.6060  49.991 1.67760
#> 40   4    7 8.3822e+00  0.01166300 -0.3334000 19.6060  49.991 1.67760
#> 41   4    9 3.8260e+00  0.02455700 -0.1799200 19.6060  49.991 1.67760
#> 42   4   11 1.7462e+00  0.00790240 -0.2506900 19.6060  49.991 1.67760
#>           ETA1        ETA2      ETA3 AGE  HT     WT SECR SEX RACE SMOK HCTZ
#> 2   -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 3   -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 4   -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 5   -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 6   -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 7   -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 8   -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 9   -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 10  -0.2677300  0.19829000 -0.163610  55 154  80.97  1.0   2    2    0    1
#> 12  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 13  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 14  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 15  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 16  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 17  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 18  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 19  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 20  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 21  -0.7096900  0.18606000  0.736900  37 179  93.21  1.2   1    1    1    0
#> 23  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 24  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 25  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 26  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 27  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 28  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 29  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 30  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 31  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 32  -0.4762000  0.20152000  0.435760  35 188  94.35  0.9   1    1    0    0
#> 34   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#> 35   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#> 36   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#> 37   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#> 38   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#> 39   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#> 40   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#> 41   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#> 42   0.0995910 -0.42932000  0.150950  67 168  74.39  0.8   2    2    0    0
#>     PROP CON OCC     DV     PRED         RES       WRES
#> 2      1   1   0  71.74 86.41800 -14.6780000 -0.1046800
#> 3      1   1   0  72.61 89.00800 -16.3980000 -0.7589100
#> 4      1   1   0  88.01 75.46800  12.5420000  1.2075000
#> 5      1   1   0  53.13 61.03600  -7.9061000 -1.5386000
#> 6      1   1   0  56.83 48.71100   8.1187000 -0.1152100
#> 7      1   1   0  51.94 38.72400  13.2160000  0.3580400
#> 8      1   1   0  52.89 30.74900  22.1410000  2.0247000
#> 9      1   1   0  26.95 19.37200   7.5781000 -1.3045000
#> 10     1   1   0  26.17 12.20200  13.9680000  1.3377000
#> 12     0   0   0 108.75 86.41800  22.3320000  0.7884900
#> 13     0   0   0  96.60 89.00800   7.5920000  0.2056300
#> 14     0   0   0  81.00 75.46800   5.5321000 -0.1610700
#> 15     0   0   0  77.07 61.03600  16.0340000  0.3367900
#> 16     0   0   0  64.57 48.71100  15.8590000 -0.4293600
#> 17     0   0   0  50.21 38.72400  11.4860000 -1.9657000
#> 18     0   0   0  64.58 30.74900  33.8310000  1.7130000
#> 19     0   0   0  50.54 24.40700  26.1330000 -0.0245510
#> 20     0   0   0  41.56 15.37500  26.1850000  0.4600400
#> 21     0   0   0  42.46  9.68460  32.7750000  5.4087000
#> 23     0   0   0   9.35  8.64180   0.7082000  0.4329000
#> 24     0   0   0   8.66  8.90080  -0.2408000 -0.2000400
#> 25     0   0   0   8.18  7.54680   0.6332100  0.2378000
#> 26     0   0   0   6.19  6.10360   0.0863880 -0.9516600
#> 27     0   0   0   7.08  4.87110   2.2089000  1.2177000
#> 28     0   0   0   4.68  3.87240   0.8076200 -1.4248000
#> 29     0   0   0   5.30  3.07490   2.2251000  0.8501900
#> 30     0   0   0   4.20  2.44070   1.7593000 -0.2344100
#> 31     0   0   0   3.92  1.53750   2.3825000  2.1772000
#> 32     0   0   0   2.75  0.96846   1.7815000  1.4402000
#> 34     0   0   0  63.15 43.20900  19.9410000  0.4473500
#> 35     0   0   0  62.41 44.50400  17.9060000  1.5121000
#> 36     0   0   0  41.05 37.73400   3.3161000  0.2789900
#> 37     0   0   0  24.31 30.51800  -6.2081000 -0.9520200
#> 38     0   0   0  18.17 24.35600  -6.1857000 -0.5277300
#> 39     0   0   0  13.23 19.36200  -6.1319000 -0.2349800
#> 40     0   0   0   8.48 15.37400  -6.8943000 -0.4056700
#> 41     0   0   0   3.92  9.68600  -5.7660000 -0.0433460
#> 42     0   0   0   1.76  6.10120  -4.3412000  0.3824000
#>  [ reached getOption("max.print") -- omitted 538 rows ]

str(simpraz.xpdb)

#> Formal class 'xpose.data' [package "xpose4"] with 8 slots
#>   ..@ Data           :'data.frame':	640 obs. of  26 variables:
#>   .. ..$ ID   : num [1:640] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ TIME : num [1:640] 0 1 2 3 4 5 6 7 9 11 ...
#>   .. ..$ IPRED: num [1:640] 0 69.2 80.2 75.3 66.9 ...
#>   .. ..$ IWRES: num [1:640] 0 0.0368 -0.0944 0.1683 -0.206 ...
#>   .. ..$ CWRES: num [1:640] 0 -0.0646 -0.9411 1.1911 -1.5154 ...
#>   .. ..$ CL   : num [1:640] 13.6 13.6 13.6 13.6 13.6 ...
#>   .. ..$ V    : num [1:640] 93.6 93.6 93.6 93.6 93.6 ...
#>   .. ..$ KA   : num [1:640] 1.22 1.22 1.22 1.22 1.22 ...
#>   .. ..$ ETA1 : num [1:640] -0.268 -0.268 -0.268 -0.268 -0.268 ...
#>   .. ..$ ETA2 : num [1:640] 0.198 0.198 0.198 0.198 0.198 ...
#>   .. ..$ ETA3 : num [1:640] -0.164 -0.164 -0.164 -0.164 -0.164 ...
#>   .. ..$ AGE  : num [1:640] 55 55 55 55 55 55 55 55 55 55 ...
#>   .. ..$ HT   : num [1:640] 154 154 154 154 154 154 154 154 154 154 ...
#>   .. ..$ WT   : num [1:640] 81 81 81 81 81 ...
#>   .. ..$ SECR : num [1:640] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ SEX  : Factor w/ 2 levels "1","2": 2 2 2 2 2 2 2 2 2 2 ...
#>   .. ..$ RACE : Factor w/ 3 levels "1","2","3": 2 2 2 2 2 2 2 2 2 2 ...
#>   .. ..$ SMOK : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ HCTZ : Factor w/ 2 levels "0","1": 2 2 2 2 2 2 2 2 2 2 ...
#>   .. ..$ PROP : Factor w/ 2 levels "0","1": 2 2 2 2 2 2 2 2 2 2 ...
#>   .. ..$ CON  : Factor w/ 2 levels "0","1": 2 2 2 2 2 2 2 2 2 2 ...
#>   .. ..$ OCC  : Factor w/ 1 level "0": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ DV   : num [1:640] 0 71.7 72.6 88 53.1 ...
#>   .. ..$ PRED : num [1:640] 0 86.4 89 75.5 61 ...
#>   .. ..$ RES  : num [1:640] 0 -14.68 -16.4 12.54 -7.91 ...
#>   .. ..$ WRES : num [1:640] 0 -0.105 -0.759 1.208 -1.539 ...
#>   ..@ SData          : NULL
#>   ..@ Data.firstonly :'data.frame':	64 obs. of  12 variables:
#>   .. ..$ SUBJECT_NO: int [1:64] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..$ ID        : int [1:64] 1 2 3 4 5 6 7 8 9 10 ...
#>   .. ..$ ETA.1.    : num [1:64] -0.2677 -0.7097 -0.4762 0.0996 -0.3529 ...
#>   .. ..$ ETA.2.    : num [1:64] 0.198 0.186 0.202 -0.429 0.098 ...
#>   .. ..$ ETA.3.    : num [1:64] -0.164 0.737 0.436 0.151 0.524 ...
#>   .. ..$ ETC.1.1.  : num [1:64] 0.00412 0.00646 0.00401 0.00321 0.00328 ...
#>   .. ..$ ETC.2.1.  : num [1:64] -0.002413 -0.002923 -0.001677 0.003449 -0.000594 ...
#>   .. ..$ ETC.2.2.  : num [1:64] 0.00971 0.00622 0.00661 0.00605 0.00691 ...
#>   .. ..$ ETC.3.1.  : num [1:64] -0.00947 -0.01828 -0.01274 -0.00658 -0.01295 ...
#>   .. ..$ ETC.3.2.  : num [1:64] 0.01757 0.00998 0.01247 0.00237 0.01117 ...
#>   .. ..$ ETC.3.3.  : num [1:64] 0.0821 0.3497 0.1956 0.0754 0.2295 ...
#>   .. ..$ OBJ       : num [1:64] 54.04 60.73 9.13 27.62 25.53 ...
#>   ..@ SData.firstonly: NULL
#>   ..@ Runno          : num 1
#>   ..@ Nsim           : NULL
#>   ..@ Doc            : NULL
#>   ..@ Prefs          :Formal class 'xpose.prefs' [package "xpose4"] with 9 slots
#>   .. .. ..@ Xvardef      :List of 14
#>   .. .. .. ..$ id        : chr "ID"
#>   .. .. .. ..$ idlab     : chr "ID"
#>   .. .. .. ..$ idv       : chr "TIME"
#>   .. .. .. ..$ occ       : chr "OCC"
#>   .. .. .. ..$ dv        : chr "DV"
#>   .. .. .. ..$ pred      : chr "PRED"
#>   .. .. .. ..$ ipred     : chr "IPRED"
#>   .. .. .. ..$ iwres     : chr "IWRES"
#>   .. .. .. ..$ wres      : chr "WRES"
#>   .. .. .. ..$ cwres     : chr "CWRES"
#>   .. .. .. ..$ res       : chr "RES"
#>   .. .. .. ..$ parms     : chr [1:6] "ETA3" "ETA2" "ETA1" "KA" ...
#>   .. .. .. ..$ covariates: chr [1:11] "SEX" "RACE" "SMOK" "HCTZ" ...
#>   .. .. .. ..$ ranpar    : chr [1:3] "ETA1" "ETA2" "ETA3"
#>   .. .. ..@ Labels       :List of 29
#>   .. .. .. ..$ OCC  : chr "Occasion"
#>   .. .. .. ..$ TIME : chr "Time"
#>   .. .. .. ..$ PRED : chr "Population predictions"
#>   .. .. .. ..$ IPRED: chr "Individual predictions"
#>   .. .. .. ..$ IPRE : chr "Individual predictions"
#>   .. .. .. ..$ WRES : chr "Weighted residuals"
#>   .. .. .. ..$ CWRES: chr "Conditional weighted residuals"
#>   .. .. .. ..$ IWRES: chr "Individual weighted residuals"
#>   .. .. .. ..$ IWRE : chr "Individual weighted residuals"
#>   .. .. .. ..$ DV   : chr "Observations"
#>   .. .. .. ..$ RES  : chr "Residuals"
#>   .. .. .. ..$ CL   : chr "Clearance"
#>   .. .. .. ..$ V    : chr "Volume"
#>   .. .. .. ..$ TAD  : chr "Time after dose"
#>   .. .. .. ..$ ID   : chr "ID"
#>   .. .. .. ..$ KA   : chr "KA"
#>   .. .. .. ..$ ETA1 : chr "ETA1"
#>   .. .. .. ..$ ETA2 : chr "ETA2"
#>   .. .. .. ..$ ETA3 : chr "ETA3"
#>   .. .. .. ..$ AGE  : chr "AGE"
#>   .. .. .. ..$ HT   : chr "HT"
#>   .. .. .. ..$ WT   : chr "WT"
#>   .. .. .. ..$ SECR : chr "SECR"
#>   .. .. .. ..$ SEX  : chr "SEX"
#>   .. .. .. ..$ RACE : chr "RACE"
#>   .. .. .. ..$ SMOK : chr "SMOK"
#>   .. .. .. ..$ HCTZ : chr "HCTZ"
#>   .. .. .. ..$ PROP : chr "PROP"
#>   .. .. .. ..$ CON  : chr "CON"
#>   .. .. ..@ Graph.prefs  :List of 102
#>   .. .. .. ..$ type     : chr "b"
#>   .. .. .. ..$ pch      : num 1
#>   .. .. .. ..$ cex      : num 0.8
#>   .. .. .. ..$ lty      : num 1
#>   .. .. .. ..$ lwd      : num 1
#>   .. .. .. ..$ col      : num 4
#>   .. .. .. ..$ fill     : chr "lightblue"
#>   .. .. .. ..$ grid     : logi FALSE
#>   .. .. .. ..$ aspect   : chr "fill"
#>   .. .. .. ..$ condvar  : NULL
#>   .. .. .. ..$ byordfun : chr "median"
#>   .. .. .. ..$ ordby    : NULL
#>   .. .. .. ..$ shingnum : num 6
#>   .. .. .. ..$ shingol  : num 0.5
#>   .. .. .. ..$ abline   : NULL
#>   .. .. .. ..$ abllwd   : num 1
#>   .. .. .. ..$ ablcol   : num 1
#>   .. .. .. ..$ abllty   : num 1
#>   .. .. .. ..$ smlwd    : num 2
#>   .. .. .. ..$ smcol    : num 2
#>   .. .. .. ..$ smlty    : num 1
#>   .. .. .. ..$ smspan   : num 0.667
#>   .. .. .. ..$ smdegr   : num 1
#>   .. .. .. ..$ lmline   : NULL
#>   .. .. .. ..$ lmlwd    : num 2
#>   .. .. .. ..$ lmcol    : num 2
#>   .. .. .. ..$ lmlty    : num 1
#>   .. .. .. ..$ suline   : NULL
#>   .. .. .. ..$ sulwd    : num 2
#>   .. .. .. ..$ sucol    : num 3
#>   .. .. .. ..$ sulty    : num 1
#>   .. .. .. ..$ suspan   : num 0.667
#>   .. .. .. ..$ sudegr   : num 1
#>   .. .. .. ..$ ids      : logi FALSE
#>   .. .. .. ..$ idsmode  : NULL
#>   .. .. .. ..$ idsext   : num 0.05
#>   .. .. .. ..$ idscex   : num 0.7
#>   .. .. .. ..$ idsdir   : chr "both"
#>   .. .. .. ..$ dilfrac  : num 0.7
#>   .. .. .. ..$ diltype  : NULL
#>   .. .. .. ..$ dilci    : num 0.95
#>   .. .. .. ..$ PIuplty  : num 2
#>   .. .. .. ..$ PIdolty  : num 2
#>   .. .. .. ..$ PImelty  : num 1
#>   .. .. .. ..$ PIuptyp  : chr "l"
#>   .. .. .. ..$ PIdotyp  : chr "l"
#>   .. .. .. ..$ PImetyp  : chr "l"
#>   .. .. .. ..$ PIupcol  : chr "black"
#>   .. .. .. ..$ PIdocol  : chr "black"
#>   .. .. .. ..$ PImecol  : chr "black"
#>   .. .. .. ..$ PIuplwd  : num 2
#>   .. .. .. ..$ PIdolwd  : num 2
#>   .. .. .. ..$ PImelwd  : num 2
#>   .. .. .. ..$ PIupltyR : num 1
#>   .. .. .. ..$ PIdoltyR : num 1
#>   .. .. .. ..$ PImeltyR : num 2
#>   .. .. .. ..$ PIuptypR : chr "l"
#>   .. .. .. ..$ PIdotypR : chr "l"
#>   .. .. .. ..$ PImetypR : chr "l"
#>   .. .. .. ..$ PIupcolR : chr "blue"
#>   .. .. .. ..$ PIdocolR : chr "blue"
#>   .. .. .. ..$ PImecolR : chr "blue"
#>   .. .. .. ..$ PIuplwdR : num 2
#>   .. .. .. ..$ PIdolwdR : num 2
#>   .. .. .. ..$ PImelwdR : num 2
#>   .. .. .. ..$ PIupltyM : num 1
#>   .. .. .. ..$ PIdoltyM : num 1
#>   .. .. .. ..$ PImeltyM : num 2
#>   .. .. .. ..$ PIuptypM : chr "l"
#>   .. .. .. ..$ PIdotypM : chr "l"
#>   .. .. .. ..$ PImetypM : chr "l"
#>   .. .. .. ..$ PIupcolM : chr "darkgreen"
#>   .. .. .. ..$ PIdocolM : chr "darkgreen"
#>   .. .. .. ..$ PImecolM : chr "darkgreen"
#>   .. .. .. ..$ PIuplwdM : num 0.5
#>   .. .. .. ..$ PIdolwdM : num 0.5
#>   .. .. .. ..$ PImelwdM : num 0.5
#>   .. .. .. ..$ PIarcol  : chr "lightgreen"
#>   .. .. .. ..$ PIlimits : num [1:2] 0.025 0.975
#>   .. .. .. ..$ bwhoriz  : logi FALSE
#>   .. .. .. ..$ bwratio  : num 1.5
#>   .. .. .. ..$ bwvarwid : logi FALSE
#>   .. .. .. ..$ bwdotpch : num 16
#>   .. .. .. ..$ bwdotcol : chr "black"
#>   .. .. .. ..$ bwdotcex : num 1
#>   .. .. .. ..$ bwreccol : chr "blue"
#>   .. .. .. ..$ bwrecfill: chr "transparent"
#>   .. .. .. ..$ bwreclty : num 1
#>   .. .. .. ..$ bwreclwd : num 1
#>   .. .. .. ..$ bwumbcol : chr "blue"
#>   .. .. .. ..$ bwumblty : num 1
#>   .. .. .. ..$ bwumblwd : num 1
#>   .. .. .. ..$ bwoutcol : chr "blue"
#>   .. .. .. ..$ bwoutcex : num 0.8
#>   .. .. .. ..$ bwoutpch : num 1
#>   .. .. .. ..$ hicol    : num 5
#>   .. .. .. ..$ hiborder : chr "black"
#>   .. .. .. ..$ hilty    : num 1
#>   .. .. .. ..$ hilwd    : num 1
#>   .. .. .. .. [list output truncated]
#>   .. .. ..@ Miss         : num -99
#>   .. .. ..@ Cat.levels   : num 4
#>   .. .. ..@ DV.Cat.levels: num 7
#>   .. .. ..@ Subset       : NULL
#>   .. .. ..@ Gam.prefs    :List of 21
#>   .. .. .. ..$ onlyfirst : logi TRUE
#>   .. .. .. ..$ wts       : logi FALSE
#>   .. .. .. ..$ start.mod : NULL
#>   .. .. .. ..$ steppit   : logi TRUE
#>   .. .. .. ..$ disp      : NULL
#>   .. .. .. ..$ nmods     : num 3
#>   .. .. .. ..$ smoother1 : num 0
#>   .. .. .. ..$ smoother2 : num 1
#>   .. .. .. ..$ smoother3 : chr "ns"
#>   .. .. .. ..$ smoother4 : chr "ns"
#>   .. .. .. ..$ arg1      : NULL
#>   .. .. .. ..$ arg2      : NULL
#>   .. .. .. ..$ arg3      : chr "df=2"
#>   .. .. .. ..$ arg4      : chr "df=3"
#>   .. .. .. ..$ excl1     : NULL
#>   .. .. .. ..$ excl2     : NULL
#>   .. .. .. ..$ excl3     : NULL
#>   .. .. .. ..$ excl4     : NULL
#>   .. .. .. ..$ extra     : NULL
#>   .. .. .. ..$ plot.ids  : logi TRUE
#>   .. .. .. ..$ medianNorm: logi TRUE
#>   .. .. ..@ Bootgam.prefs:List of 10
#>   .. .. .. ..$ n             : num 100
#>   .. .. .. ..$ algo          : chr "fluct.ratio"
#>   .. .. .. ..$ conv.value    : num 1.04
#>   .. .. .. ..$ check.interval: num 20
#>   .. .. .. ..$ start.check   : num 50
#>   .. .. .. ..$ liif          : num 0.2
#>   .. .. .. ..$ ljif.conv     : num 25
#>   .. .. .. ..$ seed          : NULL
#>   .. .. .. ..$ start.mod     : NULL
#>   .. .. .. ..$ excluded.ids  : NULL