Function to create files for the simulated Prazosin example in Xpose

Creates NONMEM data, model and output files for a model of Prazosin using simulted data.

simprazExample(overwrite = FALSE)

Arguments

overwrite
Logical. Should the function overwrite files with the same names already in the current working directory?

Details

Creates files in the current working directory named: run1.ext run1.lst run1.mod simpraz.dta xptab1

See also

Other data functions: add_transformed_columns, change_graphical_parameters, change_misc_parameters, compute.cwres, data.checkout, data_extract_or_assign, db.names, export.graph.par, export.variable.definitions, import.graph.par, import.variable.definitions, make.sb.data, nsim, par_cov_summary, read.TTE.sim.data, read.nm.tables, read_NM_output, read_nm_table, tabulate.parameters, xlabel, xpose.data, xpose.print, xpose4-package, xsubset

Examples

od = setwd(tempdir()) # move to a temp directory
(cur.files <- dir()) # current files in temp directory

#>   [1] "0050bb79487f451487bab3dd9847cb6e.png"                                                
#>   [2] "00f3ba3af1e746448d26df9e1e60660b.png"                                                
#>   [3] "032f9a70a19448a5bd2d24bb43499a64.png"                                                
#>   [4] "0638b8a3768149fe918720e5c2ef18e9.png"                                                
#>   [5] "07c8060a48ae45159f5008880b03a24d.png"                                                
#>   [6] "0aca4e05b91d422f9296466b24b5fc41.png"                                                
#>   [7] "0b5d80391b6d43c3aa6bbc29fb0dc3d6.png"                                                
#>   [8] "0b67319d692742c8834591162a3647ad.png"                                                
#>   [9] "11f19707dbfd472e8ba94d3768f7aeeb.png"                                                
#>  [10] "13eb4bd79c284311a30985e415400e6f.png"                                                
#>  [11] "1673ef29d7514fd19ee252b8a571f40f.png"                                                
#>  [12] "1840fb38b68f4899abd8e71f78a57d11.png"                                                
#>  [13] "18c045d2753b46ba8db22aef42f38bcd.png"                                                
#>  [14] "1979607b36454f3d947379ffe6e74f76.png"                                                
#>  [15] "197d0522dbcd4ad688ca8d2fecc499ba.png"                                                
#>  [16] "1c92af1409744895835a3c0d4e2c59f6.png"                                                
#>  [17] "1e28bbc7a8fe41d3913345faab932f9d.png"                                                
#>  [18] "1e4cc337309c407890aa3e987a1d4c1e.png"                                                
#>  [19] "1eb9f7f9a01044c0833b6842ba919b96.png"                                                
#>  [20] "20ba4ccd7924404d8413e0b2d0ae403a.png"                                                
#>  [21] "21757f8c08214490ad6998846c76127a.png"                                                
#>  [22] "217b1c80043a43d793764fa759c58200.png"                                                
#>  [23] "22d68fb6de3c4ab488b8ed7695644fb6.png"                                                
#>  [24] "25660a42f88b4275a0f97b6a66856b2a.png"                                                
#>  [25] "25b2bb56152d44b9a70b189bfc26c08d.png"                                                
#>  [26] "28159ee4a1f24c95b79d02665774ec23.png"                                                
#>  [27] "282110fc2f844aa7825551f1b1eda687.png"                                                
#>  [28] "28feb19f05f74b2f8a0f506418bb19ef.png"                                                
#>  [29] "29d12f915c2b40ed82991cb84980efd0.png"                                                
#>  [30] "2aab1e4536514002859576b735485971.png"                                                
#>  [31] "2af9be1ba5a347c8baf18d8211120295.png"                                                
#>  [32] "2b5b5319653042daa2f3c52deeaa7ebf.png"                                                
#>  [33] "2c5f9a30cf4a4d5385f3353ce19f8e27.png"                                                
#>  [34] "2ca138e8c45b4f49bb522521fdc77362.png"                                                
#>  [35] "2f28fd2d0501473bb17a6290901ca780.png"                                                
#>  [36] "2fce41252d4d47329913e50f0335c895.png"                                                
#>  [37] "3024d28778294651b9457c572d87dc35.png"                                                
#>  [38] "312bdd4f641749a59fdbd8ff592c60c2.png"                                                
#>  [39] "31c374d5f5cd45a4a1679294fa056b9f.png"                                                
#>  [40] "35565301aab54235a5e91cc4d8f8848a.png"                                                
#>  [41] "35840d211f7b412b852f62f42c94271f.png"                                                
#>  [42] "3695e46e124e4a8dbd8b773dec150b43.png"                                                
#>  [43] "39fedbc4fe5b4a03a8647beac6ac23cc.png"                                                
#>  [44] "3ff4926f0722486ca9ff4795a84877d6.png"                                                
#>  [45] "401a2ab6d2fd4d65a7775ce010fb0932.png"                                                
#>  [46] "403eca9f85424f18b65e3c7356ef7dc6.png"                                                
#>  [47] "42bc132d787f43278d56ce679e15ac86.png"                                                
#>  [48] "42d83b525fc84d9e9fa99780f167897d.png"                                                
#>  [49] "4597e10b718c4d3fa33307e326aebee7.png"                                                
#>  [50] "45cf4a2d43a34bfcbc57deb4b54f5628.png"                                                
#>  [51] "46e8649d8c984b2b8f63d004c0daa18e.png"                                                
#>  [52] "47677f1e4cf24fe0a8c71851db4e6d68.png"                                                
#>  [53] "48e4bb3520154c10b64299cbd8afaa47.png"                                                
#>  [54] "48fd8512eabf4ee5a81ced92f6c954f4.png"                                                
#>  [55] "4a8aa682a9f344f8b0ee9a916440d990.png"                                                
#>  [56] "4aab54725b1f4f9384cedf2de19fa2bc.png"                                                
#>  [57] "4db642e580bf4f8e8086765d3a676288.png"                                                
#>  [58] "5042bd81714f4d2981a322b267cd6b37.png"                                                
#>  [59] "51ba7ddd20f14f9abd385cbb3d550560.png"                                                
#>  [60] "51ebb745cee44ee89c85ef4ffc188de8.png"                                                
#>  [61] "5312b6b53a3b42709e45f3e277193634.png"                                                
#>  [62] "5385c84cdae240b4ba7ab0a4234c8f5e.png"                                                
#>  [63] "5451edfdbf954b7280daa7354b1db68b.png"                                                
#>  [64] "5526fe4f96a64b0b83d320af6317c785.png"                                                
#>  [65] "581fff3a38984d52bdbd385611cebf4e.png"                                                
#>  [66] "59f44cddfb284bc18a4b102eb361cf01.png"                                                
#>  [67] "5cf55312ee584409b19e7ef55e42efb0.png"                                                
#>  [68] "5d20eaffca9445ef8334872301f44db1.png"                                                
#>  [69] "5e3c7971491044f7a79e06d69ebb84cc.png"                                                
#>  [70] "5ebdec48be8b4576a60afe3881ddeac6.png"                                                
#>  [71] "5efe74ed772343ac9ea1e9c95c708d4f.png"                                                
#>  [72] "5f2ca5bf70c04883ad9b20f15dd9936a.png"                                                
#>  [73] "5f745a6bc8294270a02296cbf14e51be.png"                                                
#>  [74] "5fb8f0283aa6436d8614a17109c6b13f.png"                                                
#>  [75] "619e6a7222eb44179c1116b4605227b4.png"                                                
#>  [76] "61df093d8ad5469abb7e1aa4aa2a2c87.png"                                                
#>  [77] "63a77a84f3e143089b4098403cf88d96.png"                                                
#>  [78] "645a5415a0044341a23bfddd39ee074e.png"                                                
#>  [79] "64dba20a8b1b4b9bbe41715817fb8f77.png"                                                
#>  [80] "6719736f41f645498654ece7d6047721.png"                                                
#>  [81] "68033ae12b0d49e0b42238945b34c82c.png"                                                
#>  [82] "6a4f69c8543c4ecba44a1a4ae1ffdcee.png"                                                
#>  [83] "6bec1bdac7974acc988405b58e435aab.png"                                                
#>  [84] "6ea2f2106f8a41589089c756c3ad2aea.png"                                                
#>  [85] "6f9c5812f4964d71b1049eda02ad0bc9.png"                                                
#>  [86] "701efb73cdc3480c9a43ce8e61b33f62.png"                                                
#>  [87] "70cad80522df4008be88e80e8626692e.png"                                                
#>  [88] "712e381ea07a49358cdc9598c477500a.png"                                                
#>  [89] "7b865e52c8c74aa99fec9442c673a979.png"                                                
#>  [90] "7de295ca47944af8bea81d20c910465c.png"                                                
#>  [91] "7f536c6e88e04011a7f17b0cd093392b.png"                                                
#>  [92] "8106d1ee6f104c418dd7cd136ea079a4.png"                                                
#>  [93] "814aed0fc1094c768c7be2d413b48f0f.png"                                                
#>  [94] "81efcf41ff4d4a54a549f66edc699598.png"                                                
#>  [95] "81fc643e448c40f3a3232c35599803f7.png"                                                
#>  [96] "83c324b4b6c74c61a7d33e9fce322230.png"                                                
#>  [97] "866c9b86cbe745d0971443c1bc564f97.png"                                                
#>  [98] "878fe077c41f41ddb2ce6307f8668572.png"                                                
#>  [99] "881e5065206e4be088238cd1b3423ea1.png"                                                
#> [100] "884a0b542fb34d05bb328a3ad0d6fa0d.png"                                                
#> [101] "88bf43dc17064cca9b9180fbf8260ef7.png"                                                
#> [102] "8c7384fb7fa844f38375f55201805bba.png"                                                
#> [103] "8d1cb2598c834ea2946c5ba72702d82b.png"                                                
#> [104] "90e3ca4a34444c9991917ea2276336da.png"                                                
#> [105] "911f1e6ec49845029ba34718c4220beb.png"                                                
#> [106] "9165c672dcdc4b0b98d2138557dc7c59.png"                                                
#> [107] "93bf0f7607f144f8961edd2fe2576d5e.png"                                                
#> [108] "94b337aa422d4474a06c35c8c477c3c9.png"                                                
#> [109] "9781476cc95c4274be7a4625cc704d55.png"                                                
#> [110] "98420ce2ede7471789a3bc4a5370e7d5.png"                                                
#> [111] "98738426294b490ab6ea3c16345528d6.png"                                                
#> [112] "993e507847544181a76681b2615cdff4.png"                                                
#> [113] "9a44b7b651f541db810ffcf6c6462aa3.png"                                                
#> [114] "9b111c4d753e46c0a912975df1291073.png"                                                
#> [115] "9c9f90905a9049459a8966f1a8989e4c.png"                                                
#> [116] "9f5acd001bc14575af18627411eacee4.png"                                                
#> [117] "a047ef02268240bbb5c741503fa501b8.png"                                                
#> [118] "a36c5bcd24b9443e949c0b2f321747ca.png"                                                
#> [119] "a3d8b4f928a54974b9c88a37f32de8fc.png"                                                
#> [120] "a705660e2ea54ed7a2a37e29a5ad8871.png"                                                
#> [121] "a8fab28f868146919f2079cebd675b4b.png"                                                
#> [122] "a9ac62d23c3044b29c6e9ecabab7d045.png"                                                
#> [123] "af36fa50ccce4608a3410d694f3f179d.png"                                                
#> [124] "b2f8a6d204e84b24a1fc43b3f6795993.png"                                                
#> [125] "b3ca55f4cd444d069a88f798e580d3da.png"                                                
#> [126] "b3d75c5614ab4674afc70c7adbd13fe0.png"                                                
#> [127] "b3f46a34daa745b295848283bb9269bb.png"                                                
#> [128] "b8f9fc39316242c09b9d54f3a8f69a50.png"                                                
#> [129] "b9525313161745a2be0cf43c4711e8ab.png"                                                
#> [130] "b9d7e21acb21439e9e79db3d51e3ec78.png"                                                
#> [131] "ba2ed6adc93f49c4a380fa681807e607.png"                                                
#> [132] "ba4a3bee1fe245f18927d7758b0fce88.png"                                                
#> [133] "bbec3ac8f18d4fa1a440d9f2f4e43524.png"                                                
#> [134] "bc4fcdd8024c4e69b502d2d4b83b9146.png"                                                
#> [135] "bc8958eb5fc14689ba77a3e99c2811bf.png"                                                
#> [136] "bd6616e02fd14558b21ea64d2d3efb47.png"                                                
#> [137] "bd8f047874174c4c86b6cecb262165e8.png"                                                
#> [138] "bdd6561265b246a28a446dbaa32238e2.png"                                                
#> [139] "c4ca23a1628b48eb90957ce719539411.png"                                                
#> [140] "c714bcab2fad4bb1bbe83100018a4fe6.png"                                                
#> [141] "c89370fba7b4494e861e352680619d24.png"                                                
#> [142] "c9030fab2f4743f397820cb35dca38fb.png"                                                
#> [143] "cd32da2740634c768529c0510a6dc366.png"                                                
#> [144] "ce7ded8a887e48f788e68c07b0380905.png"                                                
#> [145] "cebcdd15195a453fbc447b68db923814.png"                                                
#> [146] "d00d9b86ebfb480ea255e407eb6692eb.png"                                                
#> [147] "d270c69c8ead4093932c20a85be0e1a6.png"                                                
#> [148] "d3fd48fc17514e44ad17d037f76890cb.png"                                                
#> [149] "d69962ca0c97475f8cfc2d18ee32875c.png"                                                
#> [150] "d803587ab0db4b23a16f47ed98a64230.png"                                                
#> [151] "d9c87dc69f524f47b6bb771ff7bcb63c.png"                                                
#> [152] "da32570d635c4cb5b413e735068d3769.png"                                                
#> [153] "da4da7c38e4c43ae8862cef9a6a9b275.png"                                                
#> [154] "da5d8f16c831458f94e0919402f98aea.png"                                                
#> [155] "dc526441975348bcb2e34310b03aa0b7.png"                                                
#> [156] "dd7d99f555ef45c4b3af470af55d01b8.png"                                                
#> [157] "dd91b6971df0441aa131201c7113544d.png"                                                
#> [158] "downloaded_packages"                                                                 
#> [159] "e050f1247f7d4a088c4daa9cf98debf7.png"                                                
#> [160] "e0a7fcc164aa4a3f8bdaa10c56b63ca3.png"                                                
#> [161] "e4061249203c4e94a13d34236184b132.png"                                                
#> [162] "e6e8657cbc4745428d5cb121f8eb690c.png"                                                
#> [163] "e7f1fbad4eba491d9c8bb633ecc2681c.png"                                                
#> [164] "e8ed7cbccaa44733a01800619b10df95.png"                                                
#> [165] "e8fe3847598949f8bab113a4ada1295a.png"                                                
#> [166] "eb699b1fd5834d738f72609a845fdd28.png"                                                
#> [167] "f10f138210d64bf49d9b004a9100f96e.png"                                                
#> [168] "f24b59fe180e4ba8a65cda862e2c638f.png"                                                
#> [169] "f2640365731b42c0a6ed7cde2b140290.png"                                                
#> [170] "f2e612bea1564ce9a36f1759d9eee964.png"                                                
#> [171] "f339d359d77b415ca062ef8dce1ce0e7.png"                                                
#> [172] "f67e1cd03d0a46df851d48675b07c1ff.png"                                                
#> [173] "f6e123e4ac8f469f912cbf60cf2334e3.png"                                                
#> [174] "f6e5c30f27ea45838fea4c9f1bd2b00c.png"                                                
#> [175] "f71b68395e154005bb47c4f98a329d98.png"                                                
#> [176] "fa989415b456485fb7ceffbafac27858.png"                                                
#> [177] "fc6c7de2125a4ac0a7b6d95376b8bbfd.png"                                                
#> [178] "feaf91a56bbf424eaca19a7dee1ae6b5.png"                                                
#> [179] "fed13b3f757b438b923b510b36dd62f9.png"                                                
#> [180] "gitmsge8d477d66aa.txt"                                                               
#> [181] "gitmsge8d495c6843.txt"                                                               
#> [182] "gitmsge8d55c20d1a.txt"                                                               
#> [183] "gitmsge8d633f2885.txt"                                                               
#> [184] "libloc_213_8f8c4eafe9e43f75.rds"                                                     
#> [185] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"                         
#> [186] "repos_https%3A%2F%2Fcran.rstudio.com%2Fbin%2Fmacosx%2Fel-capitan%2Fcontrib%2F3.4.rds"
#> [187] "repos_https%3A%2F%2Fcran.rstudio.com%2Fsrc%2Fcontrib.rds"                            

simprazExample(overwrite=TRUE) # write files

(new.files <- dir()[!(dir() %in% cur.files)])  # what files are new here?

#> [1] "rs-graphics-269e886a-a3b3-48e6-b4df-48f110b55cc9"
#> [2] "run1.ext"                                        
#> [3] "run1.lst"                                        
#> [4] "run1.mod"                                        
#> [5] "simpraz.dta"                                     
#> [6] "xptab1"                                          

file.remove(new.files) # remove these files

#> Warning: cannot remove file 'rs-graphics-269e886a-a3b3-48e6-b4df-48f110b55cc9', reason 'Directory not empty'

#> [1] FALSE  TRUE  TRUE  TRUE  TRUE  TRUE

setwd(od)  # restore working directory