Extract and set labels for Xpose data items.

This function extracts and sets label definitions in Xpose data objects.

Usage

xlabel(x, object)

xlabel(object) <- value

Arguments

x

Name of the variable to assign a label to.

object

An xpose.data object.

value

A two element vector of which the first element is the name of the variable and the second the label

Value

The label of the specified column.

Details

x should be a string exactly matching the name of a column in the data.frame in the Data slot of an xpose.data object. The name of columns defined through xpose variable definitions (see xpose.data) can be extracted using the xvardef function and to be used in the xlabel function, e.g. xlabel(xvardef("dv",object),object), which would give the label for the dv variable.

Functions

xlabel(object) <- value: sets label definitions in Xpose data objects. assigned value should be a two-element vector of which the first element is the name of the variable and the second the label

See also

Author

Niclas Jonsson

Examples

xpdb <- simpraz.xpdb

## Display label for dependent variable in the Xpose data object
xlabel("DV", xpdb)
#> [1] "Observations"

## Set label for dependent variable
xlabel(xpdb) <- c("DV", "Concentration (mg/L)")
xlabel("DV", xpdb) # how has this chnaged?
#> [1] "Concentration (mg/L)"