Correlations between covariate coefficients

This function creates a plot showing the correlations in estimates for covariate coefficients, obtained from the first step (univariate testing) in each scm performed in the bootscm.

Usage

xp.boot.par.est.corr(
  bootgam.obj = NULL,
  sd.norm = TRUE,
  by.cov.type = FALSE,
  cov.plot = NULL,
  ask.covs = FALSE,
  dotpch = 19,
  col = rgb(0.2, 0.2, 0.9, 0.75),
  ...
)

Arguments

bootgam.obj

The object created using bootscm.import(), which hold the data for plotting.

sd.norm

Perform normalization of the covariate coefficients (default is TRUE). When TRUE, the estimated covariate coefficients will be multiplied by the standard deviation of the specific covariate (both for continuous and categorical covariates).

by.cov.type

Split the plot for continuous and dichotomous covariates. Default is FALSE.

cov.plot

A character vector which lists the covariates to include in the plot. If none are specified (NULL), all covariate coefficients will be included in the plot.

ask.covs

Ask the user which covariates to include in the plot. Default is FALSE.

dotpch

The character used for plotting.

col

The colors used for plotting.

...

Additional plotting arguments may be passed to this function.

Value

No value returned.

Author

Ron Keizer

Examples

if (FALSE) { # \dontrun{
xp.boot.par.est.corr(current.bootscm, sd.norm = TRUE,
                          cov.plot = c("CLSEX", "VSEX", "CLWT"))

} # }