Inclusion frequency plot

Plot the inclusion frequencies of covariates in the final models obtained in a bootgam or bootscm. Covariates are ordered by inclusion frequency.

xp.inc.prob(bootgam.obj = NULL, boot.type = NULL, main = NULL,
  col = "#6495ED", xlb = NULL, ylb = "Covariate", ...)

Arguments

bootgam.obj
The bootgam or bootscm object.

boot.type
Either "bootgam" or "bootscm". Default is NULL, which means the user will be asked to make a choice.

main
Plot title

col
Color used for the plot.

xlb
Label for x-axis.

ylb
Label for y-axis.

…
Additional plotting parameters.

Value

A lattice plot object will be returned.