Function to ask the user for the name of a file

Asks the user for the name of a file.

xpose.ask.for.filename(object, listfile = paste("run", object@Runno, ".lst",
  sep = ""), modfile = paste("run", object@Runno, ".mod", sep = ""), ...)

Arguments

object
An xpose.data object.

listfile
A NONMEM output file

modfile
A NONMEM model file

…
Additional arguments passed to the function

Value

The name of the file if it exists, otherwise nothing is returned.

Details

Function checks if the file exists, if it does then the filename is returned from the function.