Functions to create labels for plots

Functions to create labels for plots

xpose.create.title(x, y, object, subset = NULL, funx = NULL, funy = NULL,
  no.runno = FALSE, ...)

xpose.create.label(x, object, fun, logx, autocorr.x = FALSE,
  autocorr.y = FALSE, ...)

Arguments

x
Column name for x-variable

y
Column name for y variable

object
Xpose data object

subset
Subset used for plot

funx
Function applied to x data

funy
Function applied to y data

no.runno
should we include a run number in the label

…
additional arguments passed to the function.

fun
Function applied to data

logx
is the data log transformed?

autocorr.x
Are we looking at an autocorrelation plot?

autocorr.y
Are we looking at an autocorrelation plot?

plot.text
Text of plot title

main
If "Default" then plot.text is the default title with the run number and subset expression added to the title. If the value is anything other than "Default" then that value is used as the plot title.

Value

Plot titles and labels.

Functions

xpose.create.label: Create label values