Class xpose.data

The xpose.data class is the fundamental data object in Xpose 4. It contains the data and preferences used in the creation of the Xpose plots and analyses.

Objects from the Class

Objects are most easily created by the xpose.data function, which reads the appropriate NONMEM table files and populates the slots of the object.

See also

Author

Niclas Jonsson and Andrew Hooker