Displays the Xpose license and citation information

This function displays a copy of Xpose's end user license agreement (EULA).

Usage

xpose.license.citation()

Value

The EULA.

Author

Andrew Hooker

Examples

xpose.license.citation()
#> 
#> This program is free software: you can redistribute it and/or modify
#>   it under the terms of the GNU Lesser General Public License as published by
#>   the Free Software Foundation, either version 3 of the License, or
#>   (at your option) any later version.
#>   
#>   This program is distributed in the hope that it will be useful,
#>   but WITHOUT ANY WARRANTY; without even the implied warranty of
#>   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
#>   GNU Lesser General Public License for more details.
#>   
#>   A copy of the GNU Lesser General Public License can be found in the R
#>   installation directory (/opt/R/4.5.1/lib/R/share)under licenses.  
#>   If not, see <http://www.gnu.org/licenses/>.
#> 
#> To cite xpose4 in publications, use:
#> 
#>   Jonsson, E.N. & Karlsson, M.O. (1999) Xpose--an S-PLUS based
#>   population pharmacokinetic/pharmacodynamic model building aid for
#>   NONMEM. Computer Methods and Programs in Biomedicine. 58(1):51-64.
#> 
#>   Keizer RJ, Karlsson MO, Hooker AC (2013). “Modeling and Simulation
#>   Workbench for NONMEM: Tutorial on Pirana, PsN, and Xpose.” _CPT:
#>   Pharmacometrics & Systems Pharmacology_, *2*(6).
#>   doi:10.1038/psp.2013.24 <https://doi.org/10.1038/psp.2013.24>.
#> 
#> To see these entries in BibTeX format, use 'print(<citation>,
#> bibtex=TRUE)', 'toBibtex(.)', or set
#> 'options(citation.bibtex.max=999)'.