Create and object with class "xpose.multiple.plot".

Create and object with class "xpose.multiple.plot".

xpose.multiple.plot(plotList, plotTitle = NULL, nm7 = TRUE,
  prompt = FALSE, new.first.window = FALSE, max.plots.per.page = 4,
  title = list(title.x = unit(0.5, "npc"), title.y = unit(0.5, "npc"),
  title.gp = gpar(cex = 1.2, fontface = "bold"), title.just = c("center",
  "center")), mirror = FALSE, bql.layout = FALSE, ...)

Arguments

plotList
A list of lattice plots.

plotTitle
Main title for plots.

nm7
TRUE if we are using NONMEM 7

prompt
When printing should we prompt for each new page in plot?

new.first.window
TRUE or FALSE.

max.plots.per.page
A number. Max value is 9.

title
Title properties.

mirror
Are there mirror plots in plot list?

bql.layout
Should we use layout optimized for plots with BQL (below limit of quantification) measurements?

…
Additional options passed to function.

Value

An object of class "xpose.multiple.plot".

See also

print.xpose.multiple.plot, xpose.multiple.plot.default

Other generic functions: gof, xpose4-package