Default histogram panel function for Xpose 4

This is the histogram panel function for Xpose 4. This is not intended to be ised outside the xpose.plot.histogram function. Most of the arguments take their default values from xpose.data object but this can be overridden by supplying them as argument to xpose.plot.histogram.

xpose.panel.histogram(
  x,
  object,
  breaks = NULL,
  dens = TRUE,
  hidlty = object@Prefs@Graph.prefs$hidlty,
  hidcol = object@Prefs@Graph.prefs$hidcol,
  hidlwd = object@Prefs@Graph.prefs$hidlwd,
  hiborder = object@Prefs@Graph.prefs$hiborder,
  hilty = object@Prefs@Graph.prefs$hilty,
  hicol = object@Prefs@Graph.prefs$hicol,
  hilwd = object@Prefs@Graph.prefs$hilwd,
  math.dens = NULL,
  vline = NULL,
  vllwd = 3,
  vllty = 1,
  vlcol = "grey",
  hline = NULL,
  hllwd = 3,
  hllty = 1,
  hlcol = "grey",
  bins.per.panel.equal = TRUE,
  showMean = FALSE,
  meanllwd = 3,
  meanllty = 1,
  meanlcol = "orange",
  showMedian = FALSE,
  medianllwd = 3,
  medianllty = 1,
  medianlcol = "black",
  showPCTS = FALSE,
  PCTS = c(0.025, 0.975),
  PCTSllwd = 2,
  PCTSllty = hidlty,
  PCTSlcol = "black",
  vdline = NULL,
  vdllwd = 3,
  vdllty = 1,
  vdlcol = "red",
  ...,
  groups
)

Arguments

x
Name(s) of the x-variable.

object
An xpose.data object.

breaks
The breakpoints for the histogram.

dens
Density plot on top of histogram?

hidlty
Density line type.

hidcol
Color of density line.

hidlwd
Width of density line.

hiborder
Colour of the bar borders.

hilty
Line type for the bar borders.

hicol
Fill colour for the bars.

hilwd
Width for the bar borders.

math.dens
Should a density line be drawn. Values are NULL or TRUE.

vline
NULL or a vector of locations for the vertical lines to be drawn. For example, vline=c(50,60) will draw two vertical lines. The function panel.abline is used.

vllwd
Line width of the vertical lines defined with vline. Can be a vector or a single value, for example vllwd=2 or vllwd=c(2,3).

vllty
Line type of the vertical lines defined with vline. Can be a vector or a single value, for example vllty=1 or vllty=c(1,2).

vlcol
Line color of the vertical lines defined with vline. Can be a vector or a single value, for example vlcol="red" or vllty=c("red","blue").

hline
NULL or a vector of locations for the horizontal lines to be drawn. For example, hline=c(50,60) will draw two horizontal lines. The function panel.abline is used.

hllwd
Line width of the horizontal lines defined with hline. Can be a vector or a single value, for example hllwd=2 or hllwd=c(2,3).

hllty
Line type of the horizontal lines defined with hline. Can be a vector or a single value, for example hllty=1 or hllty=c(1,2).

hlcol
Line color of the horizontal lines defined with hline. Can be a vector or a single value, for example hlcol="red" or hllty=c("red","blue").

bins.per.panel.equal
Allow for different bins in different panels for continuous data? TRUE or FALSE.

showMean
Should the mean of the data in the histogram be shown?

meanllwd
Line width of mean line.

meanllty
The line type for the mean

meanlcol
Color for the mean line

showMedian
Should the median of the data for the histogram be shown as a vertical line?

medianllwd
line width of median line.

medianllty
line type of median line.

medianlcol
color of median line.

showPCTS
Should percentiles of the data for the histogram be shown?

PCTS
A vector of percentiles to show. Can be any length.

PCTSllwd
line width of percentiles. Can be a vector of same length as PCTS.

PCTSllty
Line type of the percentiles. Can be a vector of same length as PCTS.

PCTSlcol
Color of the percentiles. Can be a vector of same length as PCTS.

vdline
vertical line different for each histogram. Must be a vector.

vdllwd
line widths

vdllty
line types

vdlcol
line colors

...
Other arguments that may be needed in the function.

groups
used to pass the conditioning variable into this function.

See also