Default QQ panel function for Xpose 4

This is the QQ panel function for Xpose 4. This is not intended to be used outside the xpose.plot.qq function. Most of the arguments take their default values from xpose.data object but this can be overridden by supplying them as argument to xpose.plot.qq.

Usage

xpose.panel.qq(
  x,
  object,
  pch = object@Prefs@Graph.prefs$pch,
  col = object@Prefs@Graph.prefs$col,
  cex = object@Prefs@Graph.prefs$cex,
  abllty = object@Prefs@Graph.prefs$abllty,
  abllwd = object@Prefs@Graph.prefs$abllwd,
  ablcol = object@Prefs@Graph.prefs$ablcol,
  grid = object@Prefs@Graph.prefs$grid,
  ...
)

Arguments

x

Name(s) of the x-variable.

object

An xpose.data object.

pch

Plot character to use.

col

Colour of lines and plot symbols.

cex

Amount to scale the plotting character by.

abllty

Line type.

abllwd

Line width.

ablcol

Line colour.

grid

logical value indicating whether a visual reference grid should be added to the graph. (Could use arguments for line type, color etc).

...

Other arguments that may be needed in the function.

See also

Author

E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins