The generic Xpose functions for box-and-whisker plots

This is a wrapper function for the lattice bwplot function.

xpose.plot.bw(
  x,
  y,
  object,
  inclZeroWRES = FALSE,
  onlyfirst = FALSE,
  samp = NULL,
  panel = xpose.panel.bw,
  groups = NULL,
  ids = FALSE,
  logy = FALSE,
  logx = FALSE,
  aspect = object@Prefs@Graph.prefs$aspect,
  funy = NULL,
  funx = NULL,
  PI = FALSE,
  by = object@Prefs@Graph.prefs$condvar,
  force.by.factor = FALSE,
  ordby = object@Prefs@Graph.prefs$ordby,
  byordfun = object@Prefs@Graph.prefs$byordfun,
  shingnum = object@Prefs@Graph.prefs$shingnum,
  shingol = object@Prefs@Graph.prefs$shingol,
  strip = function(...) strip.default(..., strip.names = c(TRUE, TRUE)),
  subset = xsubset(object),
  main = xpose.create.title(x, y, object, subset, funx, funy, ...),
  xlb = xpose.create.label(x, object, funx, logx, ...),
  ylb = xpose.create.label(y, object, funy, logy, ...),
  scales = list(),
  suline = object@Prefs@Graph.prefs$suline,
  binvar = NULL,
  bins = 10,
  mirror = FALSE,
  max.plots.per.page = 4,
  mirror.aspect = "fill",
  pass.plot.list = FALSE,
  x.cex = NULL,
  y.cex = NULL,
  main.cex = NULL,
  mirror.internal = list(strip.missing = missing(strip)),
  ...
)

Arguments

x
Name(s) of the x-variable.

y
Name(s) of the y-variable.

object
An xpose.data object.

inclZeroWRES
A logical value indicating whether rows with WRES=0 should be plotted.

onlyfirst
A logical value indicating whether only the first row per individual should be included in the plot.

samp
An integer between 1 and object@Nsim (seexpose.data-class) specifying which of the simulated data sets to extract from SData.

panel
The name of the panel function to use. This should in most cases be left as xpose.panel.bw.

groups
A string with the name of any grouping variable (used as the groups argument to panel.xyplot.

ids
A logical value indicating whether text labels should be used as plotting symbols (the variable used for these symbols indicated by the idlab Xpose data variable).

logy
Logical value indicating whether the y-axis should be logarithmic.

logx
Logical value indicating whether the x-axis should be logarithmic.

aspect
The aspect ratio of the display (see bwplot).

funy
String with the name of a function to apply to the y-variable before plotting, e.g. "abs".

funx
String with the name of a function to apply to the x-variable before plotting, e.g. "abs".

PI
Either "lines", "area" or "both" specifying whether prediction intervals (as lines, as a shaded area or both) should be computed from the data in SData and added to the display. NULL means no prediction interval.

by
A string or a vector of strings with the name(s) of the conditioning variables.

force.by.factor
Logical value. If TRUE, and by is not NULL, the variable specified by by is taken as categorical.

ordby
A string with the name of a variable to be used to reorder any factor conditioning variables (by). The variable is used in a call to the reorder function.

byordfun
The name of the function to be used when reordering a factor conditioning variable (see argument ordby).

shingnum
The number of shingles ("parts") a continuous conditioning variable should be divided into.

shingol
The amount of overlap between adjacent shingles (see argument shingnum)

strip
The name of the function to be used as the strip argument to the bwplot.

subset
A string giving the subset expression to be applied to the data before plotting. See xsubset.

main
A string giving the plot title or NULL if none.

xlb
A string giving the label for the x-axis. NULL if none.

ylb
A string giving the label for the y-axis. NULL if none.

scales
A list to be used for the scales argument in bwplot.

suline
A string giving the variable to be used to construct a smooth to superpose on the display. NULL if none. This argument is used if you want to add a superpose line of a variable not present in the y list of variables.

binvar
Variable to be used for binning.

bins
The number of bins to be used. The default is 10.

mirror
Should we create mirror plots from simulation data? Value can be FALSE, TRUE or 1 for one mirror plot, or 3 for three mirror plots.

max.plots.per.page
The maximum number of plots per page that can be created with the mirror plots.

mirror.aspect
The aspect ratio of the plots used for mirror functionality.

pass.plot.list
Should we pass the list of plots created with mirror or should we print them directly. Values can be TRUE/FALSE.

x.cex
The size of the x-axis label.

y.cex
The size of the y-axis label.

main.cex
The size of the title.

mirror.internal
an internal mirror argument used in create.mirror. Checks if the strip argument from bwplot has been used.

...
Other arguments passed to xpose.panel.bw.

See also

Author

E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins

Examples

if (FALSE) {
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Box & whisker plot of WRES vs PRED
xpose.plot.bw("WRES", "PRED", xpdb5, binvar="PRED")
}