Class "xpose.prefs"

An object of the "xpose.prefs" class holds information about all the variable and graphical preferences for a particular "xpose.data" object.

Objects from the Class

Objects can be created by calls of the form new("xpose.prefs",...) but this is usually not necessary since the "xpose.prefs" object is created at the same time as the "xpose.data" object.

See also

xvardef, xlabel, xsubset, Data, SData, xpose.data, read.nm.tables, xpose.data-class, xpose.gam