Functions to create nice looking axes when using Log scales.

The functions are used to create standard tic marks and axis labels when the axes are on the log scale.

xpose.logTicks(lim, loc = c(1, 5))

xpose.yscale.components.log10(lim, ...)

xpose.xscale.components.log10(lim, ...)

Arguments

lim
Limits

loc
Locations

...
Additional arguments passed to the function.

Details

These functions create log scales that look like they should (not the default R scales). These functions are used as input to the xscale.components argument in a lattice plot.

Functions

xpose.logTicks: Make log tic marks

xpose.xscale.components.log10: Make log scale on x-axis

See also

Examples

if (FALSE) {
xpdb5 <- xpose.data(5)
xpose.plot.default("PRED","DV",xpdb,logy=T,logx=T)
xpose.plot.default("PRED","DV",xpdb,logy=T,logx=T,
                   yscale.components = xpose.yscale.components.log10,
                   xscale.components = xpose.xscale.components.log10)

## both give the same result
}