Extract or set the value of the Subset slot.

Extract or set the value of the Subset slot of an "xpose.data" object.

Usage

xsubset(object)

xsubset(object) <- value

Arguments

object

An "xpose.data" object.

value

A string with the subset expression.

Value

A string representing the subset expression.

Details

The subset string has the same syntax as the subset argument to, e.g. panel.xyplot. Note, however, that the "xpose.data" subset is not used as an argument to panel.xyplot. It is intended as the subset argument to the Data and SData functions.

Functions

xsubset(object) <- value: assign value with a string representing the subset expression

See also

Author

Niclas Jonsson

Examples

xpdb <- simpraz.xpdb
xsubset(xpdb) <- "DV > 0"
xsubset(xpdb)
#> [1] "DV > 0"