Extracts and set Xpose variable definitions.

This function extracts and set Xpose variable definitions in "xpose.data" objects.

xvardef(x, object)

xvardef(object) <- value

Arguments

x
The name of an xpose variable (see below).

object
An xpose.data object.

value
A two element vector of which the first element is the name of the variable and the second the column name in the Data slot of the object.

Value

Returns a string with the name of the data variable defined as the Xpose data variable.

Details

The Xpose variable definitions are used to map particular variable types to column names in the data.frame in the Data slot of the "xpose.data" object. The single-valued Xpose variable definitions are: id, idlab, idv, occ, dv, pred, ipred, iwres, res. The (potentially) vector-valued Xpose variable definitions are: parms, covariates, ranpar, tvparms (parameters, covariates, random effects parameters=etas, typical value parameters). The default values of these can be found in the createXposeClasses function.

Functions

xvardef<-: reset the which column the label dv points to in the Data slot of the xpose database object

See also

xpose.data-class,xpose.prefs-class

Examples

xpdb <- simpraz.xpdb

## get the column name in the Data slot of object xpdb
## corresponding to the label dv
xvardef("dv", xpdb)

#> [1] "DV"

## reset the which column the label dv points to in the Data slot of
## object xpdb
xvardef(xpdb) <- c("dv", "DVA")