Column-transformation functions for Xpose 4

These functions transform existing Xpose 4 data columns, adding new columns.

Usage

add.absval(object, listall = TRUE, classic = FALSE)

add.dichot(object, listall = TRUE, classic = FALSE)

add.exp(object, listall = TRUE, classic = FALSE)

add.log(object, listall = TRUE, classic = FALSE)

add.tad(object, classic = FALSE)

Arguments

object

An xpose.data object.

listall

A logical operator specifying whether the items in the database should be listed.

classic

A logical operator specifying whether the function should assume the classic menu system. This is an internal option and need never be called from the command line.

Value

An xpose.data object (classic == FALSE) or null (classic == TRUE).

Details

These functions may be used to create new data columns within the Xpose data object by transforming existing ones.

Functions

add.absval(): Create a column containing the absolute values of data in another column.

add.dichot(): Create a categorical data column based on a continuous data column

add.exp(): Create an exponentiated version of an existing variable

add.log(): Create a log transformation of an existing variable

add.tad(): Create a time-after-dose (TAD) data item based upon the dose and time variables in the dataset.

See also

Author

Niclas Jonsson, Justin Wilkins and Andrew Hooker

Examples

if (FALSE) {
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Create a column containing the absolute values of data in another 
## column
add.absval(xpdb5)

## Create a categorical data column based on a continuous data column, 
## and do not list variables 
add.dichot(xpdb5, listall = FALSE)

## Create a column containing the exponentiated values of data in 
## another column
add.exp(xpdb5)

## Create a column containing log-transformations of data in another 
## column
add.log(xpdb5)

## Create a time-after-dose column
add.tad(xpdb5)
}