Import bootscm data into R/Xpose

This function imports data generated by the PsN boot_scm function into the Xpose / R environment.

Usage

bootscm.import(
  scm.folder = NULL,
  silent = FALSE,
  n.bs = NULL,
  cov.recoding = NULL,
  group.by.cov = NULL,
  skip.par.est.import = FALSE,
  dofv.forward = 3.84,
  dofv.backward = 6.64,
  runno = NULL,
  return.obj = FALSE
)

Arguments

scm.folder

The folder in which the PsN-generated bootscm data are.

silent

Don't output any progress report. Default is FALSE.

n.bs

The number of bootstraps performed. Defaults to 100.

cov.recoding

For categorical covariates that are recoded to dichotomous covariates within the bootscm configuration file, a list can be specified containing data frames for recoding. See the example below for details.

group.by.cov

Group inclusion frequencies by covariate, instead of calculating them per parameter-covariates relationship. Default is NULL, which means that the user will be asked to make a choice.

skip.par.est.import

Skip the import of all parameter estimates (in each final model in all scm's, as well as parameter estimates in first step of each scm). These data are required to make plot that show inclusion bias and correlation in parameter estimates. Importing these data takes a bit of time (may take a minute or so), so if you don't intend to make these plots anyhow this step can be skipped. Default is FALSE.

dofv.forward

dOFV value used in forward step of scm.

dofv.backward

dOFV value used in backward step of scm.

runno

The run-number of the base model for this bootSCM.

return.obj

Should the bootscm object be returned by the function?

See also

Author

Ron Keizer