Changes the label of an Xpose data item

This function allows the labels of data items in the Xpose database to be changed.

change.xlabel(object, listall = TRUE, classic = FALSE)

Arguments

object
An xpose.data object.

listall
A logical operator specifying whether the items in the database should be listed.

classic
A logical operator specifying whether the function should assume the classic menu system. This is an internal option and need never be called from the command line.

Value

An xpose.data object.

Details

This function facilitates the changing of data item labels in the object@Prefs@Labels slot.

See also

Examples

if (FALSE) {
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

xpdb5 <- change.xlabel(xpdb5)
}