Change Xpose variable definitions.

These functions allow for the changing of Xpose variable definitions like "idv" and "dv". These variable definitions are used to refer to columns of the observed data in a generic way, so that generic plotting functions can be created.

Usage

change.xvardef(
  object,
  var = ".ask",
  def = ".ask",
  listall = TRUE,
  classic = FALSE,
  check.var = FALSE,
  ...
)

change.xvardef(
  object,
  var,
  listall = FALSE,
  classic = FALSE,
  check.var = FALSE,
  ...
) <- value

Arguments

object

An xpose.data object.

var

The Xpose variable you would like to change or add to the current object. A one-element character vector (e.g. "idv"). If ".ask" the user will be prompted to input a value.

def

A vector of column names from NONMEM table files (names(object@Data)) that should be associated with this variable (e.g. c("TIME")). Multiple values are allowed. If ".ask" the user will be prompted to input values.

listall

Should the function list the database values?

classic

Is the function being used from the classic interface. This is an internal option.

check.var

Should the variables be checked against the current variables in the object?

...

Items passed to functions within this function.

value

a vector of values

Value

If called from the the command line then this function returns an xpose database. If called from the classic interface this function updates the current xpose database (.cur.db).

Functions

change.xvardef( object, var, listall = FALSE, classic = FALSE, check.var = FALSE, ... ) <- value: Change the covariate scope of the xpose database object

Additional arguments

The default xpose variables are:

id

Individual identifier column in dataset

idlab

values used for plotting ID values on data points in plots

occ

The occasion variable

dv

The dv variable

pred

The pred variable

ipred

The ipred variable

wres

The wres variable

cwres

The cwres variable

res

The res variable

parms

The parameters in the database

covariates

The covariates in the database

ranpar

The random parameters in the database

See also

Author

Andrew Hooker

Examples

## Here we load the example xpose database 
xpdb <- simpraz.xpdb

# Change the "id" variable to point to "PRED" in the xpose object
xpdb <- change.xvardef(xpdb,var="id",def="PRED")

# Check the value of the "id" variable
xvardef("id",xpdb)
#> [1] "PRED"

# Change the "idv" variable
change.xvardef(xpdb,var="idv") <- "TIME"

# Change the covariate scope
change.xvardef(xpdb,var="covariates") <- c("SEX","AGE","WT")

if (FALSE) { # \dontrun{
# Use the interactive capabilities of the function
xpdb <- change.xvardef(xpdb)
} # }