Functions changing variable definitions in Xpose 4

These functions allow customization of Xpose's graphics settings.

change.ab.graph.par(object, classic = FALSE)

change.bw.graph.par(object, classic = FALSE)

change.cond.graph.par(object, classic = FALSE)

change.dil.graph.par(object, classic = FALSE)

change.label.par(object, classic = FALSE)

change.lm.graph.par(object, classic = FALSE)

change.misc.graph.par(object, classic = FALSE)

change.pi.graph.par(object, classic = FALSE)

change.smooth.graph.par(object, classic = FALSE)

Arguments

object

An xpose.data object.

classic

A logical operator specifying whether the function should assume the classic menu system. This is an internal option and need never be called from the command line.

Value

An xpose.data object (classic == FALSE) or null (classic == TRUE).

Details

Settings can be saved and loaded using export.graph.par and import.graph.par, respectively.

Functions

change.ab.graph.par: change settings for the line of identity.

change.bw.graph.par: sets preferences for box-and-whisker plots

change.cond.graph.par: sets preferences for conditioning

change.dil.graph.par: responsible for dilution preferences

change.label.par: responsible for labelling preferences

change.lm.graph.par: responsible for linear regression lines.

change.misc.graph.par: sets basic graphics parameters, including plot type, point type and size, colour, line type, and line width.

change.pi.graph.par: responsible for prediction interval plotting preferences

change.smooth.graph.par: sets preferences for loess smooths.

See also

Author

Niclas Jonsson & Justin Wilkins

Examples

if (FALSE) {
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Change default miscellaneous graphic preferences
xpdb5 <- change.misc.graph.par(xpdb5)

## Change default linear regression line preferences, creating a new 
## object
xpdb5.a <- change.lm.graph.par(xpdb5)

## Change conditioning preferences
xpdb5 <- change.cond.graph.par(xpdb5)
}