Data functions for Xpose 4

These functions perform various tasks in managing Xpose data objects.

check.vars(vars, object, silent = FALSE)

is.readable.file(filename)

test.xpose.data(object)

xpose.bin(data, y, bins = 10)

Arguments

vars

List of variables to be checked.

object

An xpose.data object.

silent

A logical operator specifying whether output should be displayed.

filename

A filename to be checked for readability.

Value

TRUE, FALSE or NULL.

Details

These are internal Xpose functions, not intended for direct use.

See also

Author

Niclas Jonsson and Andrew Hooker