Function to create mirror plots from the generic Xpose plotting commands

This function takes the generic plotting functions from Xpose 4 and calls them multiple times with the current arguments to the functions, changing the arguments as needed for mirror plotting.

Usage

create.mirror(
  fun,
  arg.list,
  mirror,
  plotTitle,
  fix.y.limits = TRUE,
  fix.x.limits = TRUE,
  ...
)

Arguments

fun

The function name that we will call multiple times

arg.list

The arguments to that function

mirror

The type of mirror plots desired (1 or 3 mirror plots can be created)

plotTitle

The title for the plots

fix.y.limits

Should we fix all the y axes to be the same?

fix.x.limits

Should we fix all the x axes to be the same?

...

additional arguments passed to the function.

Value

a list of plots, or NULL.

Details

mostly and internal function for Xpose

See also

Author

Andrew Hooker