Exports Xpose variable definitions to a file from an Xpose data object.

This function exports variable definitions for a specified Xpose data object to a file.

export.variable.definitions(object, file = "")

Arguments

object
An xpose.data object.

file
A file name as a string.

Value

Null.

Details

This function exports variable definitions (contents of object@Prefs@Xvardef) for a given xpose.data object to a file, typically 'xpose.vardefs.ini'. Note that file format is not the same as used for graphics settings. It is a wrapper for the R function dput.

See also

import.variable.definitions, xpose.prefs-class dput

Other data functions: add_transformed_columns, change_graphical_parameters, change_misc_parameters, compute.cwres, data.checkout, data_extract_or_assign, db.names, export.graph.par, import.graph.par, import.variable.definitions, make.sb.data, nsim, par_cov_summary, read.TTE.sim.data, read.nm.tables, read_NM_output, read_nm_table, simprazExample, tabulate.parameters, xlabel, xpose.data, xpose.print, xpose4-package, xsubset

Examples

od = setwd(tempdir()) # move to a temp directory
(cur.files <- dir()) # current files in temp directory

#>  [1] "0cc2713626fd492f8859139f91364c8a.png"                       
#>  [2] "1eb8a46010e949b48853e3c926269845.png"                       
#>  [3] "1ed56d9422b2426687fb586d410fb88d.png"                       
#>  [4] "2e2821efbb94464db2282b9b0c2c5534.png"                       
#>  [5] "32530ad52d6e42b083caa830d9498c4f.png"                       
#>  [6] "3b846f1fe4af428e8c79942263c4a807.png"                       
#>  [7] "3e7674a3dbfb4f6e9584d62ed79ca368.png"                       
#>  [8] "4993faad9f9747f0b071a47d16a67506.png"                       
#>  [9] "5519cb93ae5941e480e93363a08dbdfd.png"                       
#> [10] "61333708f83d4617a0d1f929017c9661.png"                       
#> [11] "6298c54ce5594afc97ee5bdddfc88e68.png"                       
#> [12] "69d99749f40a494483568786e53e2655.png"                       
#> [13] "9651e92c41cb4b629496c51f4f58a5d9.png"                       
#> [14] "9b1e202a21f9451f8bc8cd06276127e5.png"                       
#> [15] "c33f6bb47c2e4157974bf38c89d46d0b.png"                       
#> [16] "c441ef61a9774bae9089ac4129c248df.png"                       
#> [17] "c4a8de54b3b8443db48d607ca6820999.png"                       
#> [18] "c76a464f8cf94c53ba7aed89c0c39602.png"                       
#> [19] "d7a52fa37cbf42eb91b3c7fb6ec1cff0.png"                       
#> [20] "d9fa836900ac479c9523307eda9ec140.png"                       
#> [21] "e193b8d22310445faeca56d3f5bf6d7c.png"                       
#> [22] "e1f6eb9aa2bd4c0484f6ce10a667ac1b.png"                       
#> [23] "e7894ed59b9946058594469f8c29a8f0.png"                       
#> [24] "e93973c35e9d4cb5aaaacde155dac396.png"                       
#> [25] "ee29d7e41b71444396fdae3c9e11fe68.png"                       
#> [26] "f27899702a8949d7b8110c7b684bd49b.png"                       
#> [27] "f8c664c8840141438b639e4c799c9ee7.png"                       
#> [28] "fc1044843f07434fa3e62a3d225d2dd0.png"                       
#> [29] "gitmsg195d117f2b08.txt"                                     
#> [30] "libloc_213_8f8c4eafe9e43f75.rds"                            
#> [31] "repos_https%3A%2F%2Fcran.rstudio.com%2F%2Fsrc%2Fcontrib.rds"

export.variable.definitions(simpraz.xpdb,file="xpose.vardefs.ini")
(new.files <- dir()[!(dir() %in% cur.files)])  # what files are new here?

#> [1] "xpose.vardefs.ini"

file.remove(new.files) # remove this file

#> [1] TRUE

setwd(od)  # restore working directory