Function to plot the coverage of the Numerical Predictive Check

This function takes the output from the npc command in Perl Speaks NONMEM (PsN) and makes a coverage plot. A coverage plot for the NPC looks at different prediction intervals (PIs) for each data point and calculates the total number of data points in the data set lying outside of these PIs. The plot shows the relative amount of data points outside of their PI compared to the expected amount at that PI. In addition a confidence interval around these values are computed based on the simulated data.

Usage

npc.coverage(
  npc.info = "npc_results.csv",
  main = "Default",
  main.sub = NULL,
  main.sub.cex = 0.85,
  ...
)

Arguments

npc.info

The results file from the npc command in PsN. For example, npc_results.csv, or if the file is in a separate directory ./npc_dir1/npc_results.csv.

main

A string giving the plot title or NULL if none. "Default" creates a default title.

main.sub

Used for names above each plot when using multiple plots. Should be a vector c("Group 1","Group 2")

main.sub.cex

The size of the main.sub titles.

...

Other arguments passed to xpose.multiple.plot.default, xyplot and others. Please see these functions (and below) for more descriptions of what you can do.

Value

A list of plots

Additional arguments for the NPC coverage plots

Additional plot features

CI

Specifies whether confidence intervals (as lines, a shaded area or both) should be added to the plot. Allowed values are: "area", "lines", "both", or NULL.

mark.outside.data

Should the points outside the CI be marked in a different color to identify them. Allowed values are TRUE or FALSE.

abline

Should there be a line to mark the value of y=1? Possible values are TRUE, FALSE and NULL.

Line and area control. See plot, grid.polygon and xyplot for more details.

CI.area.col

Color of the area for the CI. Defaults to "blue"

CI.area.alpha

Transparency of the CI.area.col. Defaults to 0.3.

ab.lwd

The width of the abline. Default is 1.

ab.lty

Line type of the abline. Default is "dashed"

CI.upper.lty

Line type of the line at the upper edge of the CI.

CI.upper.col

Color of the line at the upper edge of the CI.

CI.upper.lwd

The line width of the line at the upper edge of the CI.

CI.lower.lty

The line type at the lower edge of the CI.

CI.lower.col

The color of the line at the lower edge of the CI.

CI.lower.lwd

The line width of the line at the lower edge of the CI.

obs.col

The color of the observed values.

obs.pch

The type of point to use for the observed values.

obs.lty

The type of line to use for the observed values.

obs.type

The combination of lines and points to use for the observed values. Default is "b" for both.

obs.cex

The size of the points to use for the observed values.

obs.lwd

The line width to use for the observed values.

out.col

The color of the observed values that lie outside of the CI. Only used if mark.outside.data=TRUE.

out.pch

The type of point to use for the observed values that lie outside of the CI. Only used if mark.outside.data = TRUE.

out.cex

The size of the points of the observed values that lie outside of the CI. Only used if mark.outside.data = TRUE.

out.lwd

The line width of the observed values that lie outside of the CI. Only used if mark.outside.data = TRUE.

See also

Author

Andrew Hooker

Examples

if (FALSE) { # \dontrun{
library(xpose4)

npc.coverage()

## to read files in a directory different than the current working directory 
npc.file <- "./another_directory/npc_results.csv"
npc.coverage(npc.info=npc.file)
} # }