Summarize individual parameter values and covariates

These functions produce tables, printed to the screen, summarizing the individual parameter values or covariates from a dataset in Xpose 4.

cov.summary(object, onlyfirst = TRUE, subset = xsubset(object),
  inclZeroWRES = FALSE, out.file = ".screen", main = "Default",
  fill = "gray", values.to.use = xvardef("covariates", object),
  value.name = "Covariate", ...)

parm.summary(object, onlyfirst = TRUE, subset = xsubset(object),
  inclZeroWRES = FALSE, out.file = ".screen", main = "Default",
  fill = "gray", values.to.use = xvardef("parms", object),
  value.name = "Parameter", max.plots.per.page = 1, ...)

Arguments

object
An xpose.data object.

onlyfirst
Logical value indicating if only the first row per individual is included in the plot.

subset
A string giving the subset expression to be applied to the data before plotting. See xsubset.

inclZeroWRES
Logical value indicating whether rows with WRES=0 are included in the plot. The default is FALSE.

out.file
Where the results should be output to. Can be ".screen", ".ask", ".graph" or a filename in quotes.

main
The title of the plot. If "Default" then a default title is plotted. Otherwise the value should be a string like "my title" or NULL for no plot title.

fill
The color to fill the boxes in the table if the table is printed to ".graph"

values.to.use
Which values should be summarized

value.name
The name of the values

…
Other arguments passed to Data and SData.

max.plots.per.page
Maximum plots per page.

Value

Returned is the matrix of values from the table. parm.summary and cov.summary produce summaries of parameters and covariates, respectively. parm.summary produces less attractive output but supports mirror functionality. parm.summary and cov.summary utilize print.char.matrix to print the information to the screen.

Functions

cov.summary: Covariate summary

parm.summary: Parameter summary

See also

Data, SData, xpose.data-class, print.char.matrix

Other data functions: add_transformed_columns, change_graphical_parameters, change_misc_parameters, compute.cwres, data.checkout, data_extract_or_assign, db.names, export.graph.par, export.variable.definitions, import.graph.par, import.variable.definitions, make.sb.data, nsim, read.TTE.sim.data, read.nm.tables, read_NM_output, read_nm_table, simprazExample, tabulate.parameters, xlabel, xpose.data, xpose.print, xpose4-package, xsubset

Examples

## Here we load the example xpose database 
xpdb <- simpraz.xpdb

parm.summary(xpdb)

#> 
#> +----+-----------------+------+----------+----------+---------+----------------+--+
#> |    |             Mean|    SD|        Q1|    Median|       Q3|           Range| N|
#> +----+-----------------+------+----------+----------+---------+----------------+--+
#> |ETA3|    -0.0738841875|0.6318|-0.3498325| 0.0352005|0.3588275|  -1.9354-1.0009|64|
#> +----+-----------------+------+----------+----------+---------+----------------+--+
#> |ETA2|-0.00786012265625| 0.349|   -0.2862| 0.0219675|0.1990975| -0.78827-0.7406|64|
#> +----+-----------------+------+----------+----------+---------+----------------+--+
#> |ETA1|   0.007586753125|0.4507|   -0.4011|-0.0012234|  0.37004|-0.70969-0.91423|64|
#> +----+-----------------+------+----------+----------+---------+----------------+--+
#> |  KA|    1.58820765625|0.8672| 1.0169925|   1.49455| 2.065375|   0.20826-3.925|64|
#> +----+-----------------+------+----------+----------+---------+----------------+--+
#> |   V|     80.900546875| 28.77|    57.683|    78.503| 93.71575|   34.914-161.06|64|
#> +----+-----------------+------+----------+----------+---------+----------------+--+
#> |  CL|       19.8105125| 9.374|    11.884|   17.7265| 25.69625|   8.7284-44.279|64|
#> +----+-----------------+------+----------+----------+---------+----------------+--+

cov.summary(xpdb)

#> +----+--------+--+-----+
#> |    |Category|N |%    |
#> +----+--------+--+-----+
#> |SEX |1       |42|65.6 |
#> |    |2       |22|34.4 |
#> +----+--------+--+-----+
#> |RACE|1       |43|67.2 |
#> |    |2       |20|31.2 |
#> |    |3       |1 |1.6  |
#> +----+--------+--+-----+
#> |SMOK|0       |48|75.0 |
#> |    |1       |16|25.0 |
#> +----+--------+--+-----+
#> |HCTZ|0       |29|45.3 |
#> |    |1       |35|54.7 |
#> +----+--------+--+-----+
#> |PROP|0       |54|84.4 |
#> |    |1       |10|15.6 |
#> +----+--------+--+-----+
#> |CON |0       |13|20.3 |
#> |    |1       |51|79.7 |
#> +----+--------+--+-----+
#> |OCC |0       |64|100.0|
#> +----+--------+--+-----+
#> 
#> +----+----------+------+-------+------+-----+-----------+--+
#> |    |      Mean|    SD|     Q1|Median|   Q3|      Range| N|
#> +----+----------+------+-------+------+-----+-----------+--+
#> | AGE|    53.875| 10.34|   49.5|    56|   62|      24-69|64|
#> +----+----------+------+-------+------+-----+-----------+--+
#> |  HT| 171.21875| 10.38|    165|   173|  179|    140-188|64|
#> +----+----------+------+-------+------+-----+-----------+--+
#> |  WT|86.0653125| 15.54|75.9725| 85.35|96.15|51.03-136.8|64|
#> +----+----------+------+-------+------+-----+-----------+--+
#> |SECR|  1.078125|0.2012|      1|   1.1|  1.2|    0.7-1.8|64|
#> +----+----------+------+-------+------+-----+-----------+--+