Summarize individual parameter values and covariates

These functions produce tables, printed to the screen, summarizing the individual parameter values or covariates from a dataset in Xpose 4.

cov.summary(
  object,
  onlyfirst = TRUE,
  subset = xsubset(object),
  inclZeroWRES = FALSE,
  out.file = ".screen",
  main = "Default",
  fill = "gray",
  values.to.use = xvardef("covariates", object),
  value.name = "Covariate",
  ...
)

parm.summary(
  object,
  onlyfirst = TRUE,
  subset = xsubset(object),
  inclZeroWRES = FALSE,
  out.file = ".screen",
  main = "Default",
  fill = "gray",
  values.to.use = xvardef("parms", object),
  value.name = "Parameter",
  max.plots.per.page = 1,
  ...
)

Arguments

object

An xpose.data object.

onlyfirst

Logical value indicating if only the first row per individual is included in the plot.

subset

A string giving the subset expression to be applied to the data before plotting. See xsubset.

inclZeroWRES

Logical value indicating whether rows with WRES=0 are included in the plot. The default is FALSE.

out.file

Where the results should be output to. Can be ".screen", ".ask", ".graph" or a filename in quotes.

main

The title of the plot. If "Default" then a default title is plotted. Otherwise the value should be a string like "my title" or NULL for no plot title.

fill

The color to fill the boxes in the table if the table is printed to ".graph"

values.to.use

Which values should be summarized

value.name

The name of the values

...

Other arguments passed to Data and SData.

max.plots.per.page

Maximum plots per page.

Value

Returned is the matrix of values from the table. parm.summaryand cov.summary produce summaries of parameters and covariates, respectively. parm.summary produces less attractive output but supports mirror functionality. parm.summary and cov.summary utilize print.char.matrix to print the information to the screen.

Functions

cov.summary: Covariate summary

parm.summary: Parameter summary

See also

Author

Andrew Hooker & Justin Wilkins

Examples

parm.summary(simpraz.xpdb)
#> 
#> +----+-------------------+------+----------+-----------+---------+----------------+--+
#> |    |               Mean|    SD|        Q1|     Median|       Q3|           Range| N|
#> +----+-------------------+------+----------+-----------+---------+----------------+--+
#> |ETA3|    -0.073881265625|0.6318|-0.3498325|   0.035202|0.3588275|  -1.9354-1.0009|64|
#> +----+-------------------+------+----------+-----------+---------+----------------+--+
#> |ETA2|    -0.007861181875| 0.349|   -0.2862|   0.021967|0.1990975| -0.78827-0.7406|64|
#> +----+-------------------+------+----------+-----------+---------+----------------+--+
#> |ETA1|0.00758520468749997|0.4507|-0.4011075|-0.00122495|  0.37004|-0.70969-0.91423|64|
#> +----+-------------------+------+----------+-----------+---------+----------------+--+
#> |  KA|       1.5882046875|0.8672| 1.0169925|    1.49455| 2.065375|   0.20826-3.925|64|
#> +----+-------------------+------+----------+-----------+---------+----------------+--+
#> |   V|       80.900546875| 28.77|    57.683|     78.503|  93.7155|   34.914-161.06|64|
#> +----+-------------------+------+----------+-----------+---------+----------------+--+
#> |  CL|       19.810528125| 9.374|    11.884|    17.7265| 25.69625|   8.7284-44.279|64|
#> +----+-------------------+------+----------+-----------+---------+----------------+--+