Function to create a histogram of results from the randomization test tool (randtest) in PsN

Reads results from the randtest tool in PsN and then creates a histogram.

randtest.hist(results.file = "raw_results_run1.csv", df = 1, p.val = 0.05,
  main = "Default", xlim = NULL, PCTSlcol = "black", vlcol = c("red",
  "orange"), ...)

Arguments

results.file
The location of the results file from the randtest tool in PsN

df
The degrees of freedom between the full and reduced model used in the randomization test.

p.val
The p-value you would like to use.

main
The title of the plot.

xlim
The limits of the x-axis

PCTSlcol
Color of the empirical line

vlcol
Colors of the original and nominal line

…
Additional arguments that can be passed to xpose.plot.histogram, xpose.panel.histogram, histogram and other lattice-package functions.

Value

A lattice object

References

PsN

See also

xpose.plot.histogram, xpose.panel.histogram, histogram and other lattice-package functions.

Other PsN functions: boot.hist, bootscm.import, npc.coverage, read.npc.vpc.results, read.vpctab, xpose.VPC.both, xpose.VPC.categorical, xpose.VPC, xpose4-package

Examples

not_run({
  randtest.hist(results.file="randtest_dir1/raw_results_run1.csv",df=2)
})