Read the results file from a Numerical or Visual Predictive Check run in PsN

This function reads in a results file from running either the PsN command vpc or npc. The function then parses the file and passes the result to plotting functions.

Usage

read.npc.vpc.results(
  vpc.results = NULL,
  npc.results = NULL,
  verbose = FALSE,
  ...
)

Arguments

vpc.results

The name of the results file from running the PsN command vcp. Often this is named vpc_results.csv. If the file is in a directory different then the working directory then you can define a relative or absolute path to the file by, for example, ./vpc_strat_WT_4_mirror_5/vpc_results.csv.

npc.results

The name of the results file from running the PsN command npc. Often this is named npc_results.csv. relative or absolute paths to the file are allowed as for vpc.results.

verbose

Text messages passed to screen or not.

...

arguments passed to other functions.

Value

A list of values is returned.

model.file

The model file that PsN ran either the npc or vpc with

dv.var

The dependent variable used in the calculations.

idv.var

The independent variable used in the calculations. NULL if npc.results is used.

num.tables

The number of separate tables in the results file.

by.interval

The conditioning interval for the stratification variable, only returned if vpc.results is used.

result.tables

The results tables from the results file. this is a list.

Details

One of vpc.results or npc.results are necessary. If both or none are defined then the function does nothing and a NULL is returned from the function.

See also

Author

Andrew Hooker