Read NONMEM table files produced from simulation.

The function reads in NONMEM table files produced from the $SIM line in a NONMEM model file.

Usage

read_nm_table(
  nm_table,
  only_obs = FALSE,
  method = "default",
  quiet = TRUE,
  sim_num = FALSE,
  sim_name = "NSIM"
)

Arguments

nm_table

The NONMEM table file to read. A text string.

only_obs

Should the non-observation lines in the data set be removed? Currently filtered using the expected MDV column. TRUE or FALSE.

method

The methods to use for reading the tables, Can be "readr_1", "readr_2", readr_3" or "slow".

quiet

Should the error message be verbose or not?

sim_num

Should a simulation number be added to simulation tables?

sim_name

What name should one use to name the column of the simulation number?

Value

Returns a data frame of the simulated table with an added column for the simulation number. The data frame is given class c("tbl_df", "tbl", "data.frame") for easy use with dplyr.

Details

Currently the function expects the $TABLE to have a header for each new simulation. This means that the NOHEADER option or ONEHEADER option in the table file is not allowed.

See also