Function to create files for the simulated prazosin example in Xpose

Creates NONMEM data, model and output files for a model of prazosin using simulated data.

simprazExample(overwrite = FALSE)

Arguments

overwrite

Logical. Should the function overwrite files with the same names already in the current working directory?

Details

Creates files in the current working directory named: run1.ext run1.lst run1.mod simpraz.dta xptab1

See also

Author

Niclas Jonsson and Andrew Hooker

Examples

od = setwd(tempdir()) # move to a temp directory
(cur.files <- dir()) # current files in temp directory
#> [1] "downlit"          "file19df523feebd" "file19df66c8d0c7"

simprazExample(overwrite=TRUE) # write files

(new.files <- dir()[!(dir() %in% cur.files)])  # what files are new here?
#> [1] "run1.ext"    "run1.lst"    "run1.mod"    "simpraz.dta" "xptab1"     

file.remove(new.files) # remove these files
#> [1] TRUE TRUE TRUE TRUE TRUE
setwd(od)  # restore working directory