Weighted residuals (WRES) plotted against covariates, for Xpose 4

This creates a stack of plots of weighted residuals (WRES) plotted against covariates, and is a specific function in Xpose 4. It is a wrapper encapsulating arguments to the xpose.plot.default and xpose.plot.histogram functions. Most of the options take their default values from xpose.data object but may be overridden by supplying them as arguments.

wres.vs.cov(
  object,
  ylb = "WRES",
  smooth = TRUE,
  type = "p",
  main = "Default",
  ...
)

Arguments

object

An xpose.data object.

ylb

A string giving the label for the y-axis. NULL if none.

smooth

A NULL value indicates that no superposed line should be added to the graph. If TRUE then a smooth of the data will be superimposed.

type

1-character string giving the type of plot desired. The following values are possible, for details, see 'plot': '"p"' for points, '"l"' for lines, '"o"' for over-plotted points and lines, '"b"', '"c"') for (empty if '"c"') points joined by lines, '"s"' and '"S"' for stair steps and '"h"' for histogram-like vertical lines. Finally, '"n"' does not produce any points or lines.

main

The title of the plot. If "Default" then a default title is plotted. Otherwise the value should be a string like "my title" or NULL for no plot title.

...

Other arguments passed to link{xpose.plot.default} or link{xpose.plot.histogram}.

Value

Returns a stack of xyplots and histograms of CWRES versus covariates.

Details

Weighted residuals (WRES) are plotted against each covariate present, as specified in object@Prefs@Xvardef$covariates, creating a stack of plots.

A wide array of extra options controlling xyplots and histograms are available. See xpose.plot.default and xpose.plot.histogram for details.

See also

Author

E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins

Examples

if (FALSE) {
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## Here we load the example xpose database 
data(simpraz.xpdb)
xpdb <- simpraz.xpdb

## A vanilla plot
wres.vs.cov(xpdb)

## Custom colours and symbols, IDs
wres.vs.cov(xpdb, cex=0.6, pch=3, col=1, ids=TRUE)
}