Inclusion stability plot A plot of the inclusion frequency of covariates vs bootgam/bootscm iteration number. This plot can be used to evaluate whether sufficient iterations have been performed.

Inclusion stability plot

A plot of the inclusion frequency of covariates vs bootgam/bootscm iteration number. This plot can be used to evaluate whether sufficient iterations have been performed.

xp.inc.stab.cov(
  bootgam.obj = NULL,
  boot.type = NULL,
  main = NULL,
  normalize = TRUE,
  split.plots = FALSE,
  xlb = "Bootstrap replicate number",
  ylb = "Difference of estimate with final",
  ...
)

Arguments

bootgam.obj

The bootgam or bootscm object.

boot.type

Either "bootgam" or "bootscm". Default is NULL, which means the user will be asked to make a choice.

main

Plot title

normalize

Should the plot be normalized?

split.plots

Should the plots be split?

xlb

The label for the x-axis.

ylb

The label for the y-axis.

...

Additional plotting parameters

Value

A lattice plot object is returned.

See also

Author

Ron Keizer