Plot of inclusion index of covariates.

Covariate inclusion indices show the correlation in inclusion of a covariate in the final model in a bootgam or bootscm.

Usage

xp.incl.index.cov(
  bootgam.obj = NULL,
  boot.type = NULL,
  main = NULL,
  xlb = "Index",
  ylb = "Covariate",
  add.ci = FALSE,
  incl.range = NULL,
  return_plot = TRUE,
  results.tab = NULL,
  ...
)

Arguments

bootgam.obj

The bootgam or bootscm object.

boot.type

Either "bootgam" or "bootscm". Default is NULL, which means the user will be asked to make a choice.

main

Plot title.

xlb

Label for the x-axis.

ylb

Label for the y-axis.

add.ci

Add a confidence interval to the plotted data.

incl.range

Included range

return_plot

Should the function return a plot?

results.tab

Specify your own results table.

...

Additional plotting information.

Value

A lattice plot object is returned.

See also

Author

Ron Keizer