Xpose Visual Predictive Check (VPC) for both continuous and Limit of Quantification data.

Xpose Visual Predictive Check (VPC) for both continuous and Below or Above Limit of Quantification (BLQ or ALQ) data.

xpose.VPC.both(
  vpc.info = "vpc_results.csv",
  vpctab = dir(pattern = "^vpctab")[1],
  object = NULL,
  subset = NULL,
  main = "Default",
  main.sub = NULL,
  inclZeroWRES = FALSE,
  cont.logy = F,
  hline = "default",
  add.args.cont = list(),
  add.args.cat = list(),
  ...
)

Arguments

vpc.info

Name of PSN file to use. File will come from VPC command in PsN.

vpctab

Name of vpctab file produced from PsN.

object

Xpose data object.

subset

Subset of data to look at.

main

Title for plot.

main.sub

Used for names above each plot when using multiple plots. Should be a vector, e.g. c("title 1","title 2").

inclZeroWRES

Include WRES=0 rows in the computations for these plots?

cont.logy

Should the continuous plot y-axis be on the log scale?

hline

Horizontal line marking the limits of quantification. If they are defined, they must be a vector of values.

add.args.cont

Additional arguments to the continuous plot. xpose.VPC.

add.args.cat

Additional arguments to the categorical plot. xpose.VPC.categorical.

...

Additional arguments to both plots.

See also

Author

Andrew C. Hooker

Examples

if (FALSE) {
library(xpose4)

## move to the directory where results from PsN
## are found
cur.dir <- getwd()
setwd(paste(cur.dir,"/vpc_cont_LLOQ/",sep=""))

xpose.VPC()
xpose.VPC.categorical(censored=T)

xpose.VPC.both()

xpose.VPC.both(subset="DV>1.75")

xpose.VPC.both(add.args.cont=list(ylim=c(0,80)))

xpose.VPC.both(add.args.cont = list(ylim = c(0.01, 80)), xlim = c(0, 
    40), add.args.cat = list(ylim = c(0, 0.4)), cont.logy = T)

xpose.VPC.both(cont.logy=T)
}