Xpose visual predictive check for categorical data.

Xpose visual predictive check for categorical data (binary, ordered categorical and count data).

xpose.VPC.categorical(
  vpc.info = "vpc_results.csv",
  vpctab = dir(pattern = "^vpctab")[1],
  object = NULL,
  subset = NULL,
  main = "Default",
  main.sub = "Default",
  main.sub.cex = 0.85,
  real.col = 4,
  real.lty = "b",
  real.cex = 1,
  real.lwd = 1,
  median.line = FALSE,
  median.col = "darkgrey",
  median.lty = 1,
  ci.lines = FALSE,
  ci.col = "blue",
  ci.lines.col = "darkblue",
  ci.lines.lty = 3,
  xlb = "Default",
  ylb = "Proportion of Total",
  force.x.continuous = FALSE,
  level.to.plot = NULL,
  max.plots.per.page = 1,
  rug = TRUE,
  rug.col = "orange",
  censored = FALSE,
  ...
)

Arguments

vpc.info
Name of PSN file to use. File will come from VPC command in PsN.

vpctab
Name of vpctab file produced from PsN.

object
Xpose data object.

subset
Subset of data to look at.

main
Title for plot.

main.sub
Used for names above each plot when using multiple plots. Should be a vector, e.g. c("title 1","title 2").

main.sub.cex
Size of main.sub

real.col
Color of real line.

real.lty
Real line type.

real.cex
Size of real line.

real.lwd
Width of real line.

median.line
Dray a median line?

median.col
Color of median line.

median.lty
median line type.

ci.lines
Lines marking confidence interval?

ci.col
Color of CI area.

ci.lines.col
Color of CI lines.

ci.lines.lty
Type of CI lines.

xlb
X-axis label. If other than "default"" passed directly to xyplot.

ylb
Y-axis label. Passed directly to xyplot.

force.x.continuous
For the x variable to be continuous.

level.to.plot
Which levels of the variable to plot. Smallest level is 1, largest is number_of_levels. For example, with 4 levels, the largest level would be 4, the smallest would be 1.

max.plots.per.page
The number of plots per page.

rug
Should there be markings on the plot showing where the intervals for the VPC are?

rug.col
Color of the rug.

censored
Is this censored data? Censored data can be both below and above the limit of quantification.

...
Additional information passed to function.

See also

Author

Andrew C. Hooker

Examples

if (FALSE) {
library(xpose4)

## move to the directory where results from PsN
## are found
cur.dir <- getwd()
setwd(paste(cur.dir,"/binary/vpc_36",sep=""))

xpose.VPC.categorical(level.to.plot=1,max.plots.per.page=4)
xpose.VPC.categorical(level.to.plot=1,max.plots.per.page=4,by="DOSE")

## ordered categorical plots
setwd(paste(cur.dir,"/ordered_cat/vpc_45",sep=""))
xpose.VPC.categorical()

## count
setwd(paste(cur.dir,"/count/vpc65b",sep=""))
xpose.VPC.categorical()

setwd(paste(cur.dir,"/count/vpc65a",sep=""))
xpose.VPC.categorical()

}