Title

Title

Usage

xpose.bootgam(
  object,
  n = n,
  id = object@Prefs@Xvardef$id,
  oid = "OID",
  seed = NULL,
  parnam = xvardef("parms", object)[1],
  covnams = xvardef("covariates", object),
  conv.value = object@Prefs@Bootgam.prefs$conv.value,
  check.interval = as.numeric(object@Prefs@Bootgam.prefs$check.interval),
  start.check = as.numeric(object@Prefs@Bootgam.prefs$start.check),
  algo = object@Prefs@Bootgam.prefs$algo,
  start.mod = object@Prefs@Bootgam.prefs$start.mod,
  liif = as.numeric(object@Prefs@Bootgam.prefs$liif),
  ljif.conv = as.numeric(object@Prefs@Bootgam.prefs$ljif.conv),
  excluded.ids = as.numeric(object@Prefs@Bootgam.prefs$excluded.ids),
  ...
)

Arguments

object

An xpose.data object.

n

number of bootstrap iterations

id

column name of id

oid

create a new column with the original ID data

seed

random seed

parnam

ONE (and only one) model parameter name.

covnams

Covariate names to test on parameter.

conv.value

Convergence value

check.interval

How often to check the convergence

start.check

When to start checking

algo

Which algorithm to use

start.mod

which start model

liif

The liif value

ljif.conv

The convergence value for the liif

excluded.ids

ID values to exclude.

...

Used to pass arguments to more basic functions.

Value

a list of results from the bootstrap of the GAM.

See also

Examples

if (FALSE) { # \dontrun{
## filter out occasion as a covariate as only one value
all_covs <- xvardef("covariates",simpraz.xpdb)
some_covs <- all_covs[!(all_covs %in% "OCC") ] 

## here only running n=5 replicates to see that things work
##   use something like n=100 for resonable results
boot_gam_obj <- xpose.bootgam(simpraz.xpdb,5,parnam="KA",covnams=some_covs,seed=1234)
} # }