Displays the Xpose license and citation information

This function displays a copy of Xpose's end user license agreement (EULA).

xpose.license.citation()

Value

The EULA.

Examples

xpose.license.citation()

#> 
#> This program is free software: you can redistribute it and/or modify
#>   it under the terms of the GNU Lesser General Public License as published by
#>   the Free Software Foundation, either version 3 of the License, or
#>   (at your option) any later version.
#>   
#>   This program is distributed in the hope that it will be useful,
#>   but WITHOUT ANY WARRANTY; without even the implied warranty of
#>   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
#>   GNU Lesser General Public License for more details.
#>   
#>   A copy of the GNU Lesser General Public License can be found in the R
#>   installation directory (/Library/Frameworks/R.framework/Resources/share)under licenses.  
#>   If not, see <http://www.gnu.org/licenses/>.
#> 
#> 
#> To cite xpose4 in publications, use:
#> 
#>   Jonsson, E.N. & Karlsson, M.O. (1999) Xpose--an S-PLUS based
#>   population pharmacokinetic/pharmacodynamic model building aid for
#>   NONMEM. Computer Methods and Programs in Biomedicine. 58(1):51-64.
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Article{,
#>     title = {Xpose--an S-PLUS based population pharmacokinetic/pharmacodynamic model building aid for NONMEM},
#>     journal = {Computer Methods and Programs in Biomedicine},
#>     volume = {58},
#>     number = {1},
#>     pages = {51-64},
#>     year = {1999},
#>     author = {E. N. Jonsson and M. O. Karlsson},
#>   }
#>