Xpose 4 generic function for plotting multiple lattice objects on one page

Function takes a list of lattice plot objects and prints them in a multiple plot layout with a title.

xpose.multiple.plot.default(
  plotList,
  plotTitle = NULL,
  prompt = FALSE,
  new.first.window = FALSE,
  max.plots.per.page = 4,
  title = list(title.x = unit(0.5, "npc"), title.y = unit(0.5, "npc"), title.gp =
    gpar(cex = 1.2, fontface = "bold"), title.just = c("center", "center")),
  mirror = FALSE,
  bql.layout = FALSE,
  page.numbers = TRUE,
  ...
)

Arguments

plotList
A list of lattice plot objects such that plot object i can be called with plotList[[i]]

plotTitle
The title used for the multiple plot layout

prompt
If more than one page is needed do you want a prompt at the command line before the next page is printed

new.first.window
Should the first page of this plot be in the already opened window or should a new window be created

max.plots.per.page
Maximum number of plots per page in the multiple layout

title
Look of title using grid.

mirror
if the list contains mirror plots

bql.layout
should we use layout optimized for BQL measurements?

page.numbers
Should we add page numbers to multiple page plots?

...
Other arguments passed to the code in this function

Value

returns nothing

Details

Additional arguments:

title.x

Where the title should be placed in the title grid region

title.y

Where the title should be placed in the title grid region

title.just

how the title should be justified

title.gp

The par parameters for the title (see grid)

See also