Scatterplot matrix panel function for Xpose 4

This is the scatterplot matrix panel function for Xpose 4. This is not intended to be ised outside the xpose.plot.splom function. Most of the arguments take their default values from xpose.data object but this can be overridden by supplying them as argument to xpose.plot.splom.

xpose.panel.splom(
  x,
  y,
  object,
  subscripts,
  onlyfirst = TRUE,
  inclZeroWRES = FALSE,
  type = "p",
  col = object@Prefs@Graph.prefs$col,
  pch = object@Prefs@Graph.prefs$pch,
  cex = object@Prefs@Graph.prefs$cex,
  lty = object@Prefs@Graph.prefs$lty,
  lwd = object@Prefs@Graph.prefs$lwd,
  smooth = TRUE,
  smlwd = object@Prefs@Graph.prefs$smlwd,
  smlty = object@Prefs@Graph.prefs$smlty,
  smcol = object@Prefs@Graph.prefs$smcol,
  smspan = object@Prefs@Graph.prefs$smspan,
  smdegr = object@Prefs@Graph.prefs$smdegr,
  lmline = NULL,
  lmlwd = object@Prefs@Graph.prefs$lmlwd,
  lmlty = object@Prefs@Graph.prefs$lmlty,
  lmcol = object@Prefs@Graph.prefs$lmcol,
  grid = object@Prefs@Graph.prefs$grid,
  groups = NULL,
  ...
)

Arguments

x
Name(s) of the x-variable.

y
Name(s) of the y-variable.

object
An xpose.data object.

subscripts
The standard Trellis subscripts argument (see xyplot)

onlyfirst
Logical value indicating whether only the first row per individual is included in the plot.

inclZeroWRES
Logical value indicating whether rows with WRES=0 is included in the plot.

type
1-character string giving the type of plot desired. The following values are possible, for details, see 'plot': '"p"' for points, '"l"' for lines, '"o"' for over-plotted points and lines, '"b"', '"c"') for (empty if '"c"') points joined by lines, '"s"' and '"S"' for stair steps and '"h"' for histogram-like vertical lines. Finally, '"n"' does not produce any points or lines.

col
The color for lines and points. Specified as an integer or a text string. A full list is obtained by the R command colours(). The default is blue (col=4).

pch
The plotting character, or symbol, to use. Specified as an integer. See R help on points. The default is an open circle.

cex
The amount by which plotting text and symbols should be scaled relative to the default. 'NULL' and 'NA' are equivalent to '1.0'.

lty
The line type. Line types can either be specified as an integer (0=blank, 1=solid, 2=dashed, 3=dotted, 4=dotdash, 5=longdash, 6=twodash) or as one of the character strings '"blank"', '"solid"', '"dashed"', '"dotted"', '"dotdash"', '"longdash"', or '"twodash"', where '"blank"' uses 'invisible lines' (i.e., doesn't draw them).

lwd
the width for lines. Specified as an integer. The default is 1.

smooth
A NULL value indicates that no superposed line should be added to the graph. If TRUE then a smooth of the data will be superimposed.

smlwd
Line width of the x-y smooth.

smlty
Line type of the x-y smooth.

smcol
Line color of the x-y smooth.

smspan
The smoothness parameter for the x-y smooth. The default is 0.667. An argument to panel.loess.

smdegr
The degree of the polynomials to be used for the x-y smooth, up to 2. The default is 1. An argument to panel.loess.

lmline
logical variable specifying whether a linear regression line should be superimposed over an xyplot. NULL ~ FALSE. (y~x)

lmlwd
Line width of the lmline.

lmlty
Line type of the lmline.

lmcol
Line colour of the lmline.

grid
logical value indicating whether a visual reference grid should be added to the graph. (Could use arguments for line type, color etc).

groups
Name of the variable used for superpose plots.

...
Other arguments that may be needed in the function.

See also