The generic Xpose functions for QQ plots

This is a wrapper function for the lattice qqmath function.

xpose.plot.qq(
  x,
  object,
  inclZeroWRES = FALSE,
  onlyfirst = FALSE,
  samp = NULL,
  aspect = object@Prefs@Graph.prefs$aspect,
  scales = list(),
  by = object@Prefs@Graph.prefs$condvar,
  force.by.factor = FALSE,
  ordby = object@Prefs@Graph.prefs$ordby,
  byordfun = object@Prefs@Graph.prefs$byordfun,
  shingnum = object@Prefs@Graph.prefs$shingnum,
  shingol = object@Prefs@Graph.prefs$shingol,
  strip = function(...) strip.default(..., strip.names = c(TRUE, TRUE)),
  subset = xsubset(object),
  main = xpose.create.title.hist(x, object, subset, ...),
  xlb = "Quantiles of Normal",
  ylb = paste("Quantiles of ", xlabel(x, object), sep = ""),
  pch = object@Prefs@Graph.prefs$pch,
  col = object@Prefs@Graph.prefs$col,
  cex = object@Prefs@Graph.prefs$cex,
  abllty = object@Prefs@Graph.prefs$abllty,
  abllwd = object@Prefs@Graph.prefs$abllwd,
  ablcol = object@Prefs@Graph.prefs$ablcol,
  mirror = FALSE,
  max.plots.per.page = 4,
  mirror.aspect = "fill",
  pass.plot.list = FALSE,
  x.cex = NULL,
  y.cex = NULL,
  main.cex = NULL,
  mirror.internal = list(strip.missing = missing(strip)),
  ...
)

Arguments

x
A string or a vector of strings with the name(s) of the x-variable(s).

object
An "xpose.data" object.

inclZeroWRES
A logical value indicating whether rows with WRES=0 should be plotted.

onlyfirst
A logical value indicating whether only the first row per individual should be included in the plot.

samp
An integer between 1 and object@Nsim (seexpose.data-class) specifying which of the simulated data sets to extract from SData.

aspect
The aspect ratio of the display (see qqmath).

scales
A list to be used for the scales argument in qqmath.

by
A string or a vector of strings with the name(s) of the conditioning variables.

force.by.factor
Logical value. If TRUE, and by is not NULL, the variable specified by by is taken as categorical.

ordby
A string with the name of a variable to be used to reorder any factor conditioning variables (by). The variable is used in a call to the reorder function.

byordfun
The name of the function to be used when reordering a factor conditioning variable (see argument ordby).

shingnum
The number of shingles ("parts") a continuous conditioning variable should be divided into.

shingol
The amount of overlap between adjacent shingles (see argument shingnum).

strip
The name of the function to be used as the strip argument to the xyplot.

subset
A string giving the subset expression to be applied to the data before plotting. See xsubset.

main
A string giving the plot title or NULL if none.

xlb
A string giving the label for the x-axis. NULL if none.

ylb
A string giving the label for the y-axis. NULL if none.

pch
Plotting symbol.

col
Color of plotting symbol.

cex
Amount to scale the plotting character by.

abllty
Line type for qqline.

abllwd
Line width for qqline.

ablcol
Color for qqline.

mirror
Should we create mirror plots from simulation data? Value can be FALSE, TRUE or 1 for one mirror plot, or 3 for three mirror plots.

max.plots.per.page
The maximum number of plots per page that can be created with the mirror plots.

mirror.aspect
The aspect ratio of the plots used for mirror functionality.

pass.plot.list
Should we pass the list of plots created with mirror or should we print them directly. Values can be TRUE/FALSE.

x.cex
The size of the x-axis label.

y.cex
The size of the y-axis label.

main.cex
The size of the title.

mirror.internal
an internal mirror argument used in create.mirror. Checks if the strip argument from qqmath has been used.

...
Other arguments passed to xpose.plot.qq.

See also

Author

E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins

Examples

if (FALSE) {
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## A QQ plot of WRES
xpose.plot.qq("WRES", xpdb5) 
}