The Xpose 4 generic functions for scatterplot matrices.

This function is a wrapper for the lattice splom function.

Usage

xpose.plot.splom(
  plist,
  object,
  varnames = NULL,
  main = "Scatterplot Matrix",
  xlb = NULL,
  ylb = NULL,
  scales = list(),
  onlyfirst = TRUE,
  inclZeroWRES = FALSE,
  subset = xsubset(object),
  by = object@Prefs@Graph.prefs$condvar,
  force.by.factor = FALSE,
  include.cat.vars = FALSE,
  ordby = NULL,
  byordfun = object@Prefs@Graph.prefs$byordfun,
  shingnum = object@Prefs@Graph.prefs$shingnum,
  shingol = object@Prefs@Graph.prefs$shingol,
  strip = function(...) strip.default(..., strip.names = c(TRUE, TRUE)),
  groups = NULL,
  ids = object@Prefs@Graph.prefs$ids,
  smooth = TRUE,
  lmline = NULL,
  panel = xpose.panel.splom,
  aspect = object@Prefs@Graph.prefs$aspect,
  samp = NULL,
  max.plots.per.page = 4,
  mirror = FALSE,
  mirror.aspect = "fill",
  pass.plot.list = FALSE,
  x.cex = NULL,
  y.cex = NULL,
  main.cex = NULL,
  mirror.internal = list(strip.missing = missing(strip)),
  ...
)

Arguments

plist

A vector of strings containing variable names for the scatterplot matrix.

object

An "xpose.data" object.

varnames

A vector of strings containing labels for the variables in the scatterplot matrix.

main

A string giving the plot title or NULL if none.

xlb

A string giving the label for the x-axis. NULL if none.

ylb

A string giving the label for the y-axis. NULL if none.

scales

A list to be used for the scales argument in xyplot.

onlyfirst

A logical value indicating whether only the first row per individual should be included in the plot.

inclZeroWRES

A logical value indicating whether rows with WRES=0 should be plotted.

subset

A string giving the subset expression to be applied to the data before plotting. See xsubset.

by

A string or a vector of strings with the name(s) of the conditioning variables.

force.by.factor

Logical value. If TRUE, and by is not NULL, the variable specified by by is taken as categorical.

include.cat.vars

Logical value.

ordby

A string with the name of a variable to be used to reorder any factor conditioning variables (by). The variable is used in a call to the reorder.factor function.

byordfun

The name of the function to be used when reordering a factor conditioning variable (see argument ordby)

shingnum

The number of shingles ("parts") a continuous conditioning variable should be divided into.

shingol

The amount of overlap between adjacent shingles (see argument shingnum)

strip

The name of the function to be used as the strip argument to the xyplot.

groups

A string with the name of any grouping variable (used as the groups argument to panel.xyplot.

ids

A logical value indicating whether text labels should be used as plotting symbols (the variable used for these symbols indicated by the idlab xpose data variable).

smooth

A NULL value indicates that no superposed line should be added to the graph. If TRUE then a smooth of the data will be superimposed.

lmline

logical variable specifying whether a linear regression line should be superimposed over an xyplot. NULL ~ FALSE. (y~x)

panel

The name of the panel function to use.

aspect

The aspect ratio of the display (see xyplot).

samp

An integer between 1 and object@Nsim (seexpose.data-class) specifying which of the simulated data sets to extract from SData.

max.plots.per.page

The maximum number of plots per page that can be created with the mirror plots.

mirror

Should we create mirror plots from simulation data? Value can be FALSE, TRUE or 1 for one mirror plot, or 3 for three mirror plots.

mirror.aspect

The aspect ratio of the plots used for mirror functionality.

pass.plot.list

Should we pass the list of plots created with mirror or should we print them directly. Values can be TRUE/FALSE.

x.cex

The size of the x-axis label.

y.cex

The size of the y-axis label.

main.cex

The size of the title.

mirror.internal

an internal mirror argument used in create.mirror. Checks if the strip argument from qqmath has been used.

...

Other arguments passed to xpose.panel.default.

Value

Returns a scatterplot matrix graph object.

Details

If ids is TRUE, text labels are added to the plotting symbols. The labels are taken from the idlab xpose data variable. The way the text labels are plotted is governed by the idsmode argument (passed down to the panel function). idsmode=NULL (the default) means that only extreme data points are labelled while a non-NULL value adds labels to all data points (the default in Xpose 3). xpose.panel.default identifies extreme data points by fitting a loess smooth (y~x) and looking at the residuals from that fit. Points that are associated with the highest/lowest residuals are labelled. "High" and "low" are judged by the panel function parameter idsext, which gives the fraction of the total number of data points that are to be judged extreme in the "up" and "down" direction. The default value for idsext is 0.05 (see link{xpose.prefs-class}). There is also a possibility to label only the high or low extreme points. This is done through the idsdir argument to xpose.panel.default. A value of "both" (the default) means that both high and low extreme points are labelled while "up" and "down" labels the high and low extreme points respectively.

More graphical parameters may be passed to xpose.panel.splom. for example, if you want to adjust the size of the varnames and axis tick labels you can use the parameters varname.cex=0.5 and axis.text.cex=0.5.

See also

Author

E. Niclas Jonsson, Mats Karlsson, Andrew Hooker & Justin Wilkins

Examples

if (FALSE) { # \dontrun{
## xpdb5 is an Xpose data object
## We expect to find the required NONMEM run and table files for run
## 5 in the current working directory
xpdb5 <- xpose.data(5)

## CL, WT, HT, SEX with a regression line
xpose.plot.splom(c("CL", "WT", "HT", "SEX"), xpdb5, lmline = TRUE) 
} # }